% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataLake.R
\name{ABSRead}
\alias{ABSRead}
\title{Reading files from BLOB storage}
\usage{
ABSRead(Source = "", Timestamp = "", SpecificSource = F,
  FileSpecific = "")
}
\details{
DEPRECATED: See \code{\link{ReadABSFile}}
}
\keyword{internal}
