% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddDimension.R
\name{AddDimension}
\alias{AddDimension}
\title{ADD Dimension}
\usage{
AddDimension(Data, Dimension = "G", SelectHierarchy = NULL)
}
\arguments{
\item{Data}{A data chunk}

\item{Dimension}{A single character used to select the dimension to be added.
\itemize{
\item "G" For Geography
\item "S" For Segment
\item "P" For Product
}}

\item{SelectHierarchy}{The name of the hierarchy required.}
}
\value{
Returns the submitted data chunk with the additional dimensional columns added.
}
\description{
This is an internal function called by other functions to add Geog, Product or Segment
hierarchies.
}
\details{
The function requires MS Corp access to access the RDR sql database. The function will allow
the user to chose the particular hierarchy required.
}
\examples{
TAPChunks:::AddDimension(Data = Data, Dimension = "P")
}
\seealso{
Other Internal Utilities: \code{\link{AddFirmographics}},
  \code{\link{AddHierarchy}}, \code{\link{CheckChunkInfo}},
  \code{\link{CheckDemogDB}}, \code{\link{CheckMemory}},
  \code{\link{CheckPackage}},
  \code{\link{CheckRequirements}},
  \code{\link{CheckSourceYaml}},
  \code{\link{CheckSpaceDisk}}, \code{\link{ConnectToADL}},
  \code{\link{CreateTable}}, \code{\link{DetectError}},
  \code{\link{FormatList}}, \code{\link{GetConfiguration}},
  \code{\link{GetMapping}},
  \code{\link{GetTimestampByObjInternal}},
  \code{\link{GetTimestampByObj}}, \code{\link{HttpCode}},
  \code{\link{LoadDemogDB}},
  \code{\link{OpenSqlConnection}},
  \code{\link{OptionsMessage}}, \code{\link{OrderColumn}},
  \code{\link{RMarkdown}},
  \code{\link{RecreateReadADLChunkExpression}},
  \code{\link{SaveLocal}}, \code{\link{ShowADLPath}},
  \code{\link{ShowADLPrefix}},
  \code{\link{ShowConfigurationPath}},
  \code{\link{TestNet}}, \code{\link{TestVPN}},
  \code{\link{TimeStamps}}, \code{\link{Unzip}},
  \code{\link{UpdateTAPChunks}},
  \code{\link{ValidateChunk}}, \code{\link{WriteFmt}}
}
\author{
JTA - The Data Scientists
}
\concept{Internal Utilities}
\keyword{internal}
