% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Manipulation.R
\name{AddFirmographics}
\alias{AddFirmographics}
\title{Add Firmographics}
\usage{
AddFirmographics(Data, Dimension = "All")
}
\arguments{
\item{Data}{A data chunk}
}
\value{
A data chunk with additional columns
}
\description{
A standard telemetry chunk contains an org identifier
called the uuid (Universally Unique Identifier).  As a separate process
we analize the uuids and attribute certain firmographics to them.  Examples include
the country name, segment and vertical market. These attributes are stored in a
separate database having all the demographic information.  This function
helps user to easily attach the firmographic attributes to a chunk.
}
\details{
The function checks to see if the demographics database has been loaded
to the environment.If not it will first load the demographics database.

The function joins the submitted chunk to the database using the uuid and returns the
original data with the following extra columns added:
\tabular{ll}{\strong{Column}\tab\strong{Description}\cr
\strong{Clean}   \tab The clean flag to show if the org has a full set of firmographics. \cr
\strong{Country} \tab The country name as defined by the source provider.\cr
\strong{Segment} \tab The segment as calculated by our load process (typically from PC and employee sizes).\cr
\strong{Vertical}\tab The vertical market as defined by SIC codes.\cr}
}
\examples{
AddFirmographics(TestServerRolesChunk)
}
\seealso{
\code{\link{TAPChunks}}
Issue #136

Other Chunk Manipulators: \code{\link{AddOSHierarchy}},
  \code{\link{AggregateChunk}},
  \code{\link{CalculateColumn}}, \code{\link{FilterChunk}},
  \code{\link{JoinChunks}}, \code{\link{RMarkdown}},
  \code{\link{ShapeChunk}}, \code{\link{UseAsSlicer}}
}
\author{
Jonathan Tooley Associados Lda
}
