% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddGSVHierarchy.R
\name{AddGSVHierarchy}
\alias{AddGSVHierarchy}
\title{Adding Firmographics (Geo, Segment and Vertical)}
\usage{
AddGSVHierarchy(Data)
}
\arguments{
\item{Data}{A TAP data chunk that must have a uuid field.}
}
\description{
This function is a shorthand that allows you to add all three
firmographic fields at once.  The function will call the following three functions:
\enumerate{
\item{\code{\link{AddGeoHierarchy}}}
\item{\code{\link{AddSegmentHierarchy}}}
\item{\code{\link{AddVerticalHierarchy}}}
}
The functions are called without hierarchy selections and so the default hierarchy will be
added. If you wish to select the hierarchy that will be added then you may call the required
functions manually and add an option to select the hierarchy.
}
\seealso{
\code{\link{TAPChunks}}

Other RDR Integration Tools: \code{\link{AddGeoHierarchy}},
  \code{\link{AddOSHierarchy}},
  \code{\link{AddProductHierarchy}},
  \code{\link{AddSegmentHierarchy}},
  \code{\link{AddVerticalHierarchy}}
}
\author{
JTA - The Data Scientists
}
\concept{RDR Integration Tools}
