% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddHierarchy.R
\name{AddHierarchy}
\alias{AddHierarchy}
\title{Add Hierarchy}
\usage{
AddHierarchy(Data, Dimension = NULL, Hierarchy = NULL)
}
\arguments{
\item{Data}{A data chunk}

\item{Dimension}{Dimension to be added to the dataset. i.e "G" "S" "O" "P"}

\item{Hierarchy}{TODO}
}
\description{
This function allows adding all hierarchies at same time. The user can define
which dimension wants to add to the dataset.  This is an internal function that
is no longer used and should be removed in a future release.
}
\section{Reference Data Repository}{

The Reference Data Repository (RDR) allows stores master data relating to
our taxonomy.  It can be used to add taxonomy hierarchies to chunks that have had
firmographic data added. Firmographic data is
established by our data load processes (which refer to the RDR to allocate firmographics)
and is stored in a separate database.

Functions that access the RDR require an internet connection which has access to
MS Corp net.
}

\section{Geo Hierarchy}{

The function accesses the RDR and retrieves the standard geography
hierarchy that is matched to the country names that the file has.
}

\section{Segment Hierarchy}{

The function accesses the RDR and retrieves the standard geography
hierarchy that is matched to the segment names that the file has.
}

\section{Product Hierarchy}{

The raw data that we receive from SpiceWorks contains a great deal of product
information.  This is typically shown in the many product columns that we see in a typical
raw file and that gets transposed when written as a chunk.
The product descriptions are detailed and so to summarize our finindings it is necessary
to add a product hierarchy.
}

\seealso{
\code{\link{TAPChunks}}

Other Internal Utilities: \code{\link{AddDimension}},
  \code{\link{AddFirmographics}},
  \code{\link{CheckChunkInfo}}, \code{\link{CheckDemogDB}},
  \code{\link{CheckMemory}}, \code{\link{CheckPackage}},
  \code{\link{CheckRequirements}},
  \code{\link{CheckSourceYaml}},
  \code{\link{CheckSpaceDisk}}, \code{\link{ConnectToADL}},
  \code{\link{CreateTable}}, \code{\link{DetectError}},
  \code{\link{FormatList}}, \code{\link{GetConfiguration}},
  \code{\link{GetMapping}},
  \code{\link{GetTimestampByObjInternal}},
  \code{\link{GetTimestampByObj}}, \code{\link{HttpCode}},
  \code{\link{LoadDemogDB}},
  \code{\link{OpenSqlConnection}},
  \code{\link{OptionsMessage}}, \code{\link{OrderColumn}},
  \code{\link{RMarkdown}},
  \code{\link{RecreateReadADLChunkExpression}},
  \code{\link{SaveLocal}}, \code{\link{ShowADLPath}},
  \code{\link{ShowADLPrefix}},
  \code{\link{ShowConfigurationPath}},
  \code{\link{TestNet}}, \code{\link{TestVPN}},
  \code{\link{TimeStamps}}, \code{\link{Unzip}},
  \code{\link{UpdateTAPChunks}},
  \code{\link{ValidateChunk}}, \code{\link{WriteFmt}}
}
\author{
JTA - The Data Scientists
}
\concept{Internal Utilities}
\keyword{internal}
