% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Manipulation.R
\name{AddOSHierarchy}
\alias{AddOSHierarchy}
\title{Add OS Mappings to a data chunk}
\usage{
AddOSHierarchy(Data, OSColumn, Format)
}
\arguments{
\item{Data}{This is a required field and is usually a TAP chunk but it can be any item of class "data.table" or "data.frame".
If the user submits a data frame then this will first be converted to a data table before the calculated column is added.
This is because data tables are far more efficient in R than a data frame.
The resulting table that is returned by this function will also be a data table even if a data frame was submitted.}

\item{OSColumn}{The column name to be mapped. If none, it has as default 'device_os' or 'os_name'}

\item{Format}{'Full' to show all OS mappings | 'Unknown' to show only OS that are unknown | 'NULL' to not create a report.}
}
\value{
Returns a data table without the original column of the os and a new one named 'os.map'
}
\description{
This function reads Operating System mappings from the RDR. The RDR typically
stores text patterns which are used to match to the os field in the raw data.

This makes it easy for the user to deal with inconsistencies and to get a simple
and consistent hierarchy for OS.
This function is mainly used for Workload and PCclient data.
}
\section{OS Hierarchy}{

The function will read the OS name from the data chunk and will match to the pattern
that is stored in the RDR.  After replacing all the names matched it will show the
percentage of OS values that were not able to be recognised. The original column
denoting operating system is not returned and the following columns are added:
\tabular{ll}{
\strong{Platform} \tab Server or Client \cr
\strong{Architecture} \tab \cr
\strong{Ecosystem} \tab \cr
\strong{OS} \tab }
}

\section{AddGeoHierarchy Reference Data Repository}{
NA
}

\section{Reference Data Repository}{

The Reference Data Repository (RDR) allows stores master data relating to
our taxonomy.  It can be used to add taxonomy hierarchies to chunks that have had
firmographic data added (See \code{\link{AddFirmographics}}). Firmographic data is
established by our data load processes (which refer to the RDR to allocate firmographics)
and is stored in a separate database.

Typically functions using the RDR will try and use firmographic data in the chunk
submitted but if they can't detect firmographic data they will attempt to add this by calling
AddFirmographics for you.

Functions that access the RDR require an internet connection which has access to
MS Corp net.
}

\examples{
AddOSHierarchy(TestClientChunk)
}
\seealso{
\code{\link{TAPChunks}}
Issue #116

Other Chunk Manipulators: \code{\link{AddFirmographics}},
  \code{\link{AggregateChunk}},
  \code{\link{CalculateColumn}}, \code{\link{FilterChunk}},
  \code{\link{JoinChunks}}, \code{\link{RMarkdown}},
  \code{\link{ShapeChunk}}, \code{\link{UseAsSlicer}}

Other RDR Integration Tools: \code{\link{AddGeoHierarchy}},
  \code{\link{AddHierarchy}},
  \code{\link{AddProductHierarchy}},
  \code{\link{AddSegmentHierarchy}},
  \code{\link{AddVerticalHierarchy}}
}
\author{
Jonathan Tooley Associados Lda
}
\keyword{internal}
