% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddSegmentHierarchy.R
\name{AddSegmentHierarchy}
\alias{AddSegmentHierarchy}
\title{Add Segment hierarchy to a data chunk}
\usage{
AddSegmentHierarchy(Data, SelectHierarchy = NULL)
}
\arguments{
\item{Data}{This is a required field and is usually a TAP chunk but it can be any item of class "data.table" or "data.frame".
If the user submits a data frame then this will first be converted to a data table before the calculated column is added.
This is because data tables are far more efficient in R than a data frame.
The resulting table that is returned by this function will also be a data table even if a data frame was submitted.}

\item{SelectHierarchy}{Name of the Hierarchy to be added in the Chunk}
}
\description{
This function allows you to add a Segment hierarchy to the chunk.
}
\section{Segment Hierarchy}{

The function accesses the RDR and retrieves the standard geography
hierarchy that is matched to the segment names that the file has.
}

\section{Hierarchy available}{

\tabular{rllll}{
\strong{Hierarchy Name}   \tab \strong{Level1} \tab \strong{Level2} \tab \strong{Level3} \tab \strong{Level4}\cr
\strong{Org Segment (default)}      \tab Sector          \tab SubSector       \tab Segment         \tab Segment_Code}
}

\section{Columns fields description}{

\tabular{ll}{
\strong{Sector}      \tab Consumer, Commercial or Public\cr
\strong{SubSector}   \tab An aggregation of the segment\cr
\strong{Segment}     \tab The segment name\cr
\strong{Segment_Code}\tab A three character mnemonic}
}

\section{AddGeoHierarchy Reference Data Repository}{
NA
}

\examples{
AddSegmentHierarchy(TestEmailChunk)
}
\seealso{
\code{\link{TAPChunks}}

Other RDR Integration Tools: \code{\link{AddGSVHierarchy}},
  \code{\link{AddGeoHierarchy}},
  \code{\link{AddOSHierarchy}},
  \code{\link{AddProductHierarchy}},
  \code{\link{AddVerticalHierarchy}}
}
\author{
JTA - The Data Scientists
}
\concept{RDR Integration Tools}
