% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddVerticalHierarchy.R
\name{AddVerticalHierarchy}
\alias{AddVerticalHierarchy}
\title{Add Vertical hierarchy to a data chunk}
\usage{
AddVerticalHierarchy(Data, SelectHierarchy = NULL)
}
\arguments{
\item{Data}{This is a required field and is usually a TAP chunk but it can be any item of class "data.table" or "data.frame".
If the user submits a data frame then this will first be converted to a data table before the calculated column is added.
This is because data tables are far more efficient in R than a data frame.
The resulting table that is returned by this function will also be a data table even if a data frame was submitted.}

\item{SelectHierarchy}{Name of the Hierarchy to be added in the Chunk}
}
\description{
This function allows you to add a Vertical hierarchy to the chunk.
}
\section{Vertical Hierarchy}{

The function accesses the RDR and retrieves the standard vertical
hierarchy that is matched to the vertical names that the file has.
}

\section{Hierarchy available}{

\tabular{llll}{
\strong{Hierarchy Name}         \tab \strong{Level1} \tab \strong{Level2} \tab \strong{Level3}    \cr
\strong{Org Vertical (default)} \tab Vertical        \tab Industry        \tab IndustrySector}
}

\section{Reference Data Repository}{

The Reference Data Repository (RDR) allows stores master data relating to
our taxonomy.  It can be used to add taxonomy hierarchies to chunks that have had
firmographic data added. Firmographic data is
established by our data load processes (which refer to the RDR to allocate firmographics)
and is stored in a separate database.

Functions that access the RDR require an internet connection which has access to
MS Corp net.
}

\examples{
AddVerticalHierarchy(TestEmailChunk)
}
\seealso{
\code{\link{TAPChunks}}

Other RDR Integration Tools: \code{\link{AddGSVHierarchy}},
  \code{\link{AddGeoHierarchy}},
  \code{\link{AddOSHierarchy}},
  \code{\link{AddProductHierarchy}},
  \code{\link{AddSegmentHierarchy}}
}
\author{
JTA - The Data Scientists
}
\concept{RDR Integration Tools}
