% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataLake.R
\name{ADLDirectory}
\alias{ADLDirectory}
\title{List files in the data lake}
\usage{
ADLDirectory(Directory = "", Info = F)
}
\details{
This function has been renamed and is deprecated. See \code{\link{ListADLDirectory}}
}
\keyword{internal}
