% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShowADLPath.R
\name{ADLPath}
\alias{ADLPath}
\title{Finding the path to files in the data lake}
\usage{
ADLPath(Source = "")
}
\details{
This has been renamed and is deprecated. See \code{\link{ShowADLPath}}
}
\seealso{
\code{\link{TAPChunks}}
}
\keyword{internal}
