% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShowADLPrefix.R
\name{ADLPrefix}
\alias{ADLPrefix}
\title{Find the filename prefix associated with a source}
\usage{
ADLPrefix(Source)
}
\details{
This is deprecated and has bee renamed now to ShowADLPrefix. See \code{\link{ShowADLPrefix}}
}
\seealso{
\code{\link{TAPChunks}}
}
\keyword{internal}
