% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataLake.R
\name{ADLRead}
\alias{ADLRead}
\title{Reading files from the Azure Data Lake}
\usage{
ADLRead(SpecificPath, FileName, SaveChunk = T, ShowInfo = T)
}
\details{
This function has been renamed. See \code{\link{ReadADLFile}}
}
\keyword{internal}
