% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadADLChunk.R
\name{ADLReadChunk}
\alias{ADLReadChunk}
\title{Reading source chunks from the Azure Data Lake}
\usage{
ADLReadChunk()
}
\details{
This function has been renamed.  See \code{\link{ReadADLChunk}}
}
\keyword{internal}
