% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ApplyLongitudinalSet.R
\name{ApplyLongitudinalSet}
\alias{ApplyLongitudinalSet}
\title{Forming the Longitudinal Data Set}
\usage{
ApplyLongitudinalSet(Data)
}
\arguments{
\item{Data}{A data chunk that must have an org identifier and a timestamp column.}
}
\value{
The data chunk with just the organization in the longitudinal cohort.
}
\description{
This function considers each organisation identifier that exists in
the submitted chunk and looks to see the number of distinct time periods in which
each organization appears. Then the data is filtered so that any organizations that
don't have data for all of the periods in the submitted chunk are removed.
The resulting chunk will have the same cohort of organizations in each period.
}
\examples{
ApplyLongitudinalSet(TestEmailChunk)
}
\seealso{
\code{\link{TAPChunks}}

Other Weighting and cleaning functions: \code{\link{ApplyWeightCalc}},
  \code{\link{CleanChunk}},
  \code{\link{ShowLongitudinalSize}},
  \code{\link{ShowWeightCalc}}
}
\author{
JTA - The Data Scientists
}
\concept{Weighting and cleaning functions}
