% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ApplyPenetration.R
\name{ApplyPenetration}
\alias{ApplyPenetration}
\title{Apply/Calculate penetration in Data Set}
\usage{
ApplyPenetration(Data, Dimension, PenetrationDimension)
}
\arguments{
\item{Data}{Dataset where the penetration will be applied}

\item{Dimension}{Dimensions to be included in the final output}

\item{PenetrationDimension}{Dimension that is the Penetration target}
}
\description{
Apply/Calculate penetration in Data Set
}
\seealso{
Other Reporting and saving tools: \code{\link{AppendCSV}},
  \code{\link{ExportToSQL}}, \code{\link{PublishADL}},
  \code{\link{SaveADL}}, \code{\link{UseAsSlicer}}
}
\author{
JTA - The Data Scientists
}
\concept{Reporting and saving tools}
