% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BeginIterate.R
\name{BeginIterate}
\alias{BeginIterate}
\title{Delimiter for use when setting up Iterations}
\usage{
BeginIterate()
}
\description{
TAP scripts can be run in the usual fashion from within your development environment
or under the control of a compute tool, such as the IterateScript function.  In order to be run
under iteration a script must use deliminators to show where an iterate block begins and where it
ends.  The BeginIterate function marks the start of an iterate block and must always be paired with
an \code{\link{EndIterate}} function, which will show where the block ends.
}
\section{The Iterate Control Block}{

When this function is called it will first check to see if an iterate control block has been
established.  If not, then this is created in the user's environment that called the BeginIterate
function.  The environment is named \strong{Iter_env} and is used to store the objects that exist
in the user's environment when the BeginIterate function is called.
}

\section{Asset List}{

The function takes an inventory of objects that exist in the user's environment and this inventory
is stored in the Iterate Control Block. This will allow the \code{\link{EndIterate}} function to
clean the objects created within the block to keep memory requirements under control.
}

\seealso{
\code{\link{TAPChunks}}

Other Compute Tools: \code{\link{CleanMemory}},
  \code{\link{EndIterate}}, \code{\link{IterateScript}},
  \code{\link{Protect}}
}
\author{
JTA - The Data Scientists
}
\concept{Compute Tools}
