% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoldText.R
\name{BoldText}
\alias{BoldText}
\title{BoldText}
\usage{
BoldText(Text)
}
\arguments{
\item{Text}{String to apply bold codes}
}
\value{
Returns a string preceded by Esc[1m and closed with Esc[22m
}
\description{
Apply underline ANSI codes arount the submitted string
}
\examples{
  cat(BoldText("Hello World!"))
message(BoldText("Status100"))
}
\seealso{
Other Text Formatters: \code{\link{BlueText}},
  \code{\link{GreenText}}, \code{\link{RedText}},
  \code{\link{UnderLineText}}, \code{\link{YellowText}},
  \code{\link{catInfo}}
}
\author{
JTA - The Data Scientists
}
\keyword{internal}
