% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ListCacheDirectory.R
\name{CacheDirectory}
\alias{CacheDirectory}
\title{List files in the local cache}
\usage{
CacheDirectory(Source = "", ShowAttributes = FALSE)
}
\details{
This function is deprecated and has been renamed.  See \code{\link{ListCacheDirectory}}
}
\keyword{internal}
