% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CacheControls.R
\name{CachePath}
\alias{CachePath}
\title{Establishing and pointing to the cache}
\usage{
CachePath(Directory)
}
\details{
This function has been renamed.  See \code{\link{ShowCachePath}}
}
\keyword{internal}
