% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateColumn.R
\name{CalculatedColumn}
\alias{CalculatedColumn}
\title{Add Calculated Columns to data chunks}
\usage{
CalculatedColumn(Data, Filter = 1 == 1, Calculation = .SD, By = NULL, ...)
}
\description{
This has been deprecated as it has now been replaced by \code{\link{CalculateColumn}}
}
\seealso{
\code{\link{TAPChunks}}
}
\author{
JTA - The Data Scientists
}
