% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Configuration.R
\name{catError}
\alias{catError}
\title{catError}
\usage{
catError(Text, Highlight = NULL)
}
\arguments{
\item{Text}{TODO}

\item{Highlight}{TODO}
}
\description{
Print Error message with Highlighed text
}
