% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catStyle.R
\name{catStyle}
\alias{catStyle}
\title{Defining Message Style}
\usage{
catStyle(Header, Text, Style, Color = "black")
}
\arguments{
\item{Header}{String to define the chapter}

\item{Text}{A string with the message that we wish to report to the user}

\item{Style}{The initial of the style pretended. (b-bold, i-italic, u-underline, s-strike)}

\item{Color}{The color to be aplied.  (red, green, yellow, blue)}
}
\description{
This function allows us to customize the console messages by applying styles.
}
\examples{
 TAPChunks:::catStyle(Header="Header:", Text="Text", Style = "ui", Color = "blue")
}
\seealso{
\code{\link{TAPChunks}}

Other Console Messaging Functions: \code{\link{catError}},
  \code{\link{catInfo}}, \code{\link{catSuccess}},
  \code{\link{catSuperUser}}, \code{\link{catWarning}}
}
\author{
JTA - The Data Scientists
}
\concept{Console Messaging Functions}
\keyword{internal}
