% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catSuccess.R
\name{catSuccess}
\alias{catSuccess}
\title{Print Success Messages}
\usage{
catSuccess(Text, Highlight = NULL)
}
\arguments{
\item{Text}{A string with the message that we wish to report to the user}

\item{Highlight}{A string that can be found in the string that was passed to the Text parameter
and which will be highlighted in bold text.
Defaults to NULL which results in no highlighting.}
}
\description{
Print Success message with Highlighed text.
}
\seealso{
\code{\link{TAPChunks}}

Other Console Messaging Functions: \code{\link{catError}},
  \code{\link{catInfo}}, \code{\link{catStyle}},
  \code{\link{catSuperUser}}, \code{\link{catWarning}}
}
\author{
JTA - The Data Scientists
}
\concept{Console Messaging Functions}
\keyword{internal}
