% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Configuration.R
\name{catWarning}
\alias{catWarning}
\title{catWarn}
\usage{
catWarning(Text, Highlight = NULL)
}
\arguments{
\item{Text}{TODO}

\item{Highlight}{TODO}
}
\description{
Print Warning message with Highlighed text
}
