% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckChunkInfo.R
\name{CheckChunkInfo}
\alias{CheckChunkInfo}
\title{Check Chunk Files information}
\usage{
CheckChunkInfo(Source = "ALL", From = "", To = "")
}
\arguments{
\item{Source}{TODO}

\item{From}{The timestamp of the first month of data.  If this is not provided this defaults
to 2016M01 but in general the user should select a timestamp.}

\item{To}{The timestamp of the last month of data needed when the user needs to get a range
of timestamps.  If this is left blank then the system returns just the single timestamp provided
in the to parameter.}
}
\description{
TODO
}
\seealso{
\code{\link{TAPChunks}}

Other Internal Utilities: \code{\link{AddDimension}},
  \code{\link{AddFirmographics}},
  \code{\link{AddHierarchy}}, \code{\link{CheckDemogDB}},
  \code{\link{CheckMemory}}, \code{\link{CheckPackage}},
  \code{\link{CheckRequirements}},
  \code{\link{CheckSourceYaml}},
  \code{\link{CheckSpaceDisk}}, \code{\link{ConnectToADL}},
  \code{\link{CreateTable}}, \code{\link{DetectError}},
  \code{\link{FormatList}}, \code{\link{GetConfiguration}},
  \code{\link{GetMapping}},
  \code{\link{GetTimestampByObjInternal}},
  \code{\link{GetTimestampByObj}}, \code{\link{HttpCode}},
  \code{\link{LoadDemogDB}},
  \code{\link{OpenSqlConnection}},
  \code{\link{OptionsMessage}}, \code{\link{OrderColumn}},
  \code{\link{RMarkdown}},
  \code{\link{RecreateReadADLChunkExpression}},
  \code{\link{SaveLocal}}, \code{\link{ShowADLPath}},
  \code{\link{ShowADLPrefix}},
  \code{\link{ShowConfigurationPath}},
  \code{\link{TestNet}}, \code{\link{TestVPN}},
  \code{\link{TimeStamps}}, \code{\link{Unzip}},
  \code{\link{UpdateTAPChunks}},
  \code{\link{ValidateChunk}}, \code{\link{WriteFmt}}
}
\author{
JTA - The Data Scientists
}
\concept{Internal Utilities}
\keyword{internal}
