% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InternalUtilities.R
\name{CheckDemogDB}
\alias{CheckDemogDB}
\title{Testing that the demographics database is loaded}
\usage{
CheckDemogDB()
}
\value{
Returns a logical TRUE if the database is loaded and FALSE if not.
}
\description{
The TAP system uses a database of demographics to perform a variety of tasks:
 Filter Unclean orgs from datasets
 Compute distributions of firmographics to calculate weights
 Add firmographic data to other datasets to be used as slicers
 Define hierarchies in reports

This function checks that the system has the database loaded.
}
\details{
The database is stored in a separate environment called TAP_env.
The function tests that this environment exists and that it has the
following three objects:
 demog_helper  : This is a data table with all of the known Spiceworks organizations
 target_helper : This is a data table with the population targets that are provided by the Market Model
 unclean       : This is a vector of unclean organisation UUIDs

If the environment does not exist it calls LoadDemogDB() to establish the environment
and returns the status of that operation.
}
\seealso{
\code{\link{TAPChunks}}

Other Internal Utilities: \code{\link{AppendCSV}},
  \code{\link{CheckChunkInfo}}, \code{\link{CheckMemory}},
  \code{\link{CheckRequirements}},
  \code{\link{CheckSourceYaml}},
  \code{\link{CheckSpaceDisk}}, \code{\link{ConnectToADL}},
  \code{\link{FormatList}}, \code{\link{GetConfiguration}},
  \code{\link{GetMapping}}, \code{\link{HttpCode}},
  \code{\link{LoadDemogDB}}, \code{\link{OptionsMessage}},
  \code{\link{OrderColumn}}, \code{\link{SaveLocal}},
  \code{\link{TestNet}}, \code{\link{TestVPN}},
  \code{\link{TimeStamps}}, \code{\link{Unzip}},
  \code{\link{ValidateChunk}}
}
\author{
Jonathan Tooley Associados Lda
}
\keyword{internal}
