% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InternalUtilities.R
\name{CheckMemory}
\alias{CheckMemory}
\title{Check Memory}
\usage{
CheckMemory(Request = NULL)
}
\arguments{
\item{Request}{chunk files required}
}
\description{
This function checks the available memory and estimates the
memory required to open the chunk files.
}
\details{
The function will lookup the size of the file requested and compare to
the memory available.  If the memory is not sufficient the function will switch
to an interactive mode where it will offer options to clear the memory.
}
\examples{
CheckMemory("SWWorkloadWL2017M12.Rdata")
}
\seealso{
\code{\link{TAPChunks}}
#174

Other Internal Utilities: \code{\link{AppendCSV}},
  \code{\link{CheckChunkInfo}}, \code{\link{CheckDemogDB}},
  \code{\link{CheckRequirements}},
  \code{\link{CheckSourceYaml}},
  \code{\link{CheckSpaceDisk}}, \code{\link{ConnectToADL}},
  \code{\link{FormatList}}, \code{\link{GetConfiguration}},
  \code{\link{GetMapping}}, \code{\link{HttpCode}},
  \code{\link{LoadDemogDB}}, \code{\link{OptionsMessage}},
  \code{\link{OrderColumn}}, \code{\link{SaveLocal}},
  \code{\link{TestNet}}, \code{\link{TestVPN}},
  \code{\link{TimeStamps}}, \code{\link{Unzip}},
  \code{\link{ValidateChunk}}
}
\author{
Jonathan Tooley Associados Lda
}
