% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InternalUtilities.R
\name{CheckSpaceDisk}
\alias{CheckSpaceDisk}
\title{Check Disk}
\usage{
CheckSpaceDisk(Request)
}
\arguments{
\item{Request}{chunk files required}
}
\description{
TODO
TODO
}
\details{
TODO
TODO
}
\examples{
CheckSpaceDisk("SWWorkloadWL2017M12.Rdata")
}
\seealso{
\code{\link{TAPChunks}}

Other Internal Utilities: \code{\link{AppendCSV}},
  \code{\link{CheckChunkInfo}}, \code{\link{CheckDemogDB}},
  \code{\link{CheckMemory}},
  \code{\link{CheckRequirements}},
  \code{\link{CheckSourceYaml}},
  \code{\link{ConnectToADL}}, \code{\link{FormatList}},
  \code{\link{GetConfiguration}}, \code{\link{GetMapping}},
  \code{\link{HttpCode}}, \code{\link{LoadDemogDB}},
  \code{\link{OptionsMessage}}, \code{\link{OrderColumn}},
  \code{\link{SaveLocal}}, \code{\link{TestNet}},
  \code{\link{TestVPN}}, \code{\link{TimeStamps}},
  \code{\link{Unzip}}, \code{\link{ValidateChunk}}
}
\author{
Jonathan Tooley Associados Lda
}
