% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Manipulation.R
\name{CleanChunk}
\alias{CleanChunk}
\title{Stripping Spiceworks chunks of unclean records}
\usage{
CleanChunk(Data)
}
\arguments{
\item{Data}{Chunk to be cleaned}
}
\value{
Returns a data table containing only the rows in the original file that are
associated with clean organizations. If the routine was not able to run then the
original data set is returned unaltered.
}
\description{
When we load Spiceworks demographics data we allocate taxonomy data
for Geography, Segment and Vertical.  When an org identifier (UUID) has all three
of these it is marked as being clean.  This function allows us to easily remove
all the orgs that are not clean from a data set.
}
\details{
The function has a dependency in that the demographics data base needs to be loaded
to the R environment. A user may check this by running CheckDemogDB().  The demographics
data base is established by running LoadDemogDB ().

When this function is called it first validates that it has been sent a valid chunk.
Next it uses the uuid column to locate all rows in the input file that have an org identifier
that is listed as not being clean in the demographics database.  These rows are
filtered from the data before it is returned.
}
\examples{
 email <- CleanChunk(TestEmailChunk)
}
\seealso{
\code{\link{TAPChunks}}

Other Weighting and cleaning functions: \code{\link{ApplyWeightCalc}},
  \code{\link{ShowWeightCalc}}
}
\author{
Jonathan Tooley Associados Lda
}
