% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConnectToADL.R
\name{ConnectToADL}
\alias{ConnectToADL}
\title{Making a connection to the data lake}
\usage{
ConnectToADL(AdlClientId = NULL, AdlClientSecret = NULL,
  AdlTenantId = NULL)
}
\arguments{
\item{AdlClientId}{The client identifier}

\item{AdlClientSecret}{The client secret}

\item{AdlTenantId}{The tenant identifier}
}
\value{
Returns an access token (or NULL on error)
}
\description{
This function allows us to establish a connection to the Data Lake.  It is intended to
be used internally by other functions in this package.  If there is no network access or
the system cannot access the Data Lake the function returns NULL.  If a connection can be
made it returns an access token.
}
\examples{
TAPChunks:::ConnectToADL(AdlClientId     = "e20ebcde-ec2d-47ab-805a-9b157638bf04",
                                   AdlClientSecret = "9LDqQfc0U+4+S7JsosPQ3HD9eLVvw9PiQYQEXU55y00=",
                                   AdlTenantId     = "72f988bf-86f1-41af-91ab-2d7cd011db47")
}
\seealso{
\code{\link{TAPChunks}}

Other Internal Utilities: \code{\link{AddDimension}},
  \code{\link{AddFirmographics}},
  \code{\link{AddHierarchy}}, \code{\link{CheckChunkInfo}},
  \code{\link{CheckDemogDB}}, \code{\link{CheckMemory}},
  \code{\link{CheckPackage}},
  \code{\link{CheckRequirements}},
  \code{\link{CheckSourceYaml}},
  \code{\link{CheckSpaceDisk}}, \code{\link{CreateTable}},
  \code{\link{DetectError}}, \code{\link{FormatList}},
  \code{\link{GetConfiguration}}, \code{\link{GetMapping}},
  \code{\link{GetTimestampByObjInternal}},
  \code{\link{GetTimestampByObj}}, \code{\link{HttpCode}},
  \code{\link{LoadDemogDB}},
  \code{\link{OpenSqlConnection}},
  \code{\link{OptionsMessage}}, \code{\link{OrderColumn}},
  \code{\link{RMarkdown}},
  \code{\link{RecreateReadADLChunkExpression}},
  \code{\link{SaveLocal}}, \code{\link{ShowADLPath}},
  \code{\link{ShowADLPrefix}},
  \code{\link{ShowConfigurationPath}},
  \code{\link{TestNet}}, \code{\link{TestVPN}},
  \code{\link{TimeStamps}}, \code{\link{Unzip}},
  \code{\link{UpdateTAPChunks}},
  \code{\link{ValidateChunk}}, \code{\link{WriteFmt}}
}
\author{
JTA - The Data Scientists
}
\concept{Internal Utilities}
\keyword{internal}
