% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConvertTime.R
\name{ConvertTime}
\alias{ConvertTime}
\title{Format seconds into a string with minutes and seconds}
\usage{
ConvertTime(TimeSec)
}
\arguments{
\item{TimeSec}{A numeric value containing the number of seconds to be converted.}
}
\description{
This routine accepts a numeric value, which can be passed as a numeric value or as a string,
and will return a formatted string.
}
\examples{
TAPChunks:::ConvertTime(25)
TAPChunks:::ConvertTime(3600)
}
\seealso{
\code{\link{TAPChunks}}
}
\author{
JTA - The Data Scientists
}
\concept{Internal Functions}
\keyword{internal}
