% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InternalUtilities.R
\name{CreateTable}
\alias{CreateTable}
\title{ConvertTime (Internal Function used in ExportToSQL)}
\usage{
CreateTable(Connection, InputData, DbSchema, DbTableName)
}
\arguments{
\item{Connection}{TODO}

\item{InputData}{TODO}

\item{DbSchema}{TODO}

\item{DbTableName}{TODO}
}
\description{
TODO
}
\examples{
TODO
}
\seealso{
\code{\link{TAPChunks}}
}
\author{
Jonathan Tooley Associados Lda
}
