% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CacheControls.R
\name{DeleteCache}
\alias{DeleteCache}
\title{Deleting files from the local cache}
\usage{
DeleteCache(PatternFile = "**", ForceDelete = F)
}
\arguments{
\item{PatternFile}{Pattern name of the files to be deleted. Default is "*".}

\item{ForceDelete}{If True will also delete folders.}
}
\value{
Returns an error status to show if the command completed succesfully or not
}
\description{
Users will need to maintain their local cache. This function allows all the cache
to be cleared or just selected files.
}
\details{
The filename can either be an explicit name or can use
wildcards.  DeleteCache("*") will remove all files. The function
will confirm the choice before actually removing files.
}
\examples{
DeleteCache("*")
}
\seealso{
\code{\link{TAPChunks}}

Other Cache Controls: \code{\link{CleanMemory}},
  \code{\link{DeleteOutdatedCache}},
  \code{\link{ListCacheDirectory}}, \code{\link{Protect}},
  \code{\link{RefreshCache}}, \code{\link{ShowCachePath}},
  \code{\link{ShowCacheStatus}},
  \code{\link{ShowConfigurationPath}},
  \code{\link{Unprotect}}
}
\author{
Jonathan Tooley Associados Lda
}
