% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeleteOutdatedCache.R
\name{DeleteOutdatedCache}
\alias{DeleteOutdatedCache}
\title{Delete outdated Rdata files in cache folder}
\usage{
DeleteOutdatedCache()
}
\description{
Delete Rdata files with more 30 days in cache folder
}
\examples{
DeleteOutdatedCache()
}
\seealso{
\code{\link{TAPChunks}}

Other Cache Controls: \code{\link{DeleteCache}},
  \code{\link{ListCacheDirectory}},
  \code{\link{RefreshCache}}, \code{\link{ShowCachePath}},
  \code{\link{ShowCacheStatus}}, \code{\link{Unprotect}}
}
\author{
JTA - The Data Scientists
}
\concept{Cache Controls}
