% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Document.R
\name{doc_content}
\alias{doc_content}
\title{Documenting a Data File's Content}
\usage{
doc_content(df)
}
\arguments{
\item{df}{a Data Table to be documented}
}
\description{
A function to write the Roxygen code to add an example of a file's content
}
\details{
When run this function will take the top 9 rows of the data table passed
as a parameter and convert the remailing content into a table using roxygen syntax.
}
\examples{
cat(doc_content(TestEmailChunk))
}
\seealso{
Other Documentation Support Tools: \code{\link{doc_cols}},
  \code{\link{doc_variables}}, \code{\link{static_help}}
}
\author{
JTA Associados Lda
}
\keyword{internal}
