% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DocumentFileColumns.R
\name{DocumentFileColumns}
\alias{DocumentFileColumns}
\title{Documenting a data file's columns}
\usage{
DocumentFileColumns(df)
}
\arguments{
\item{df}{a Data Table}
}
\description{
A function to write the Roxygen code to describe the columns in a data file.
}
\details{
This function will produce text that can be copied into the roxygen comments
section of a datafile description.  The resulting text will add a table to the documentation with
two columns: Column Name and Description.  The first column is then filled with a list of
all of the columns in the data file that was passed to this function.
}
\examples{
cat(DocumentFileColumns(TestEmailChunk))
}
\seealso{
Other Documentation Support Tools: \code{\link{DocumentFileContent}},
  \code{\link{DocumentFileVariable}},
  \code{\link{DocumentHTML}}
}
\author{
JTA - The Data Scientists
}
\concept{Documentation Support Tools}
\keyword{internal}
