% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DocumentFileContent.R
\name{DocumentFileContent}
\alias{DocumentFileContent}
\title{Documenting a Data File's Content}
\usage{
DocumentFileContent(df)
}
\arguments{
\item{df}{a Data Table to be documented}
}
\description{
A function to write the Roxygen code to add an example of a file's content
}
\details{
When run this function will take the top 9 rows of the data table passed
as a parameter and convert the remailing content into a table using roxygen syntax.
}
\examples{
cat(doc_content(TestEmailChunk))
}
\seealso{
Other Documentation Support Tools: \code{\link{DocumentFileColumns}},
  \code{\link{DocumentFileVariable}},
  \code{\link{DocumentHTML}}
}
\author{
JTA - The Data Scientists
}
\concept{Documentation Support Tools}
\keyword{internal}
