% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DocumentFileVariable.R
\name{DocumentFileVariable}
\alias{DocumentFileVariable}
\title{Documenting a TAP Chunk's variables}
\usage{
DocumentFileVariable(df)
}
\arguments{
\item{df}{a TAP Chunk (Data Table with a variable column)}
}
\description{
A function to write the Roxygen code to describe the Variables in a Chunk
}
\details{
This function will produce text that can be copied into the roxgen comments
of a data file description.  It will produce a table listing all the possible values in the variable
column of the chunk passed to the function.
}
\examples{
cat(doc_variables(TestEmailChunk))
}
\seealso{
Other Documentation Support Tools: \code{\link{DocumentFileColumns}},
  \code{\link{DocumentFileContent}},
  \code{\link{DocumentHTML}}
}
\author{
JTA - The Data Scientists
}
\concept{Documentation Support Tools}
\keyword{internal}
