% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evens_only.R
\name{evens_only}
\alias{evens_only}
\title{Return Evens}
\usage{
evens_only(x)
}
\arguments{
\item{x}{vector of numeric values}
}
\value{
even values
}
\description{
Given a vector of numbers, this will allow the user to see the even numbers only
}
