% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example}
\alias{example}
\title{Example dataset to be included in a Package}
\format{
A data frame with columns:
\describe{
\item{x}{normally distributed predictor.}
\item{y}{Outcome variable obtained with the relation \eqn{y = 2x + \epsilon}.}
}
}
\source{
Data generated internally.
}
\usage{
example
}
\description{
Example dataset to be included in a Package
}
\keyword{datasets}
