% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Configuration.R
\name{GetConfiguration}
\alias{GetConfiguration}
\title{Get configuration settings}
\usage{
GetConfiguration(Directory = NULL)
}
\arguments{
\item{Directory}{Directory where Configuration files are stored}
}
\description{
Get the file configurations like Keys to access ADL and ABS and also the source YAML file
}
\examples{
GetConfiguration()
}
\seealso{
\code{\link{TAPChunks}}

Other Internal Utilities: \code{\link{AppendCSV}},
  \code{\link{CheckChunkInfo}}, \code{\link{CheckDemogDB}},
  \code{\link{CheckMemory}},
  \code{\link{CheckRequirements}},
  \code{\link{CheckSourceYaml}},
  \code{\link{CheckSpaceDisk}}, \code{\link{ConnectToADL}},
  \code{\link{FormatList}}, \code{\link{GetMapping}},
  \code{\link{HttpCode}}, \code{\link{LoadDemogDB}},
  \code{\link{OptionsMessage}}, \code{\link{OrderColumn}},
  \code{\link{SaveLocal}}, \code{\link{TestNet}},
  \code{\link{TestVPN}}, \code{\link{TimeStamps}},
  \code{\link{Unzip}}, \code{\link{ValidateChunk}}
}
\author{
Jonathan Tooley Associados Lda
}
