% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InternalUtilities.R
\name{GetMapping}
\alias{GetMapping}
\title{Get Mapping table from RDR}
\usage{
GetMapping(Dimension = NULL)
}
\arguments{
\item{Dimension}{name the dimension}
}
\value{
The mapping table
}
\description{
This function returns the Mapping table from RDR
}
\seealso{
\code{\link{TAPChunks}}

Other Internal Utilities: \code{\link{AppendCSV}},
  \code{\link{CheckChunkInfo}}, \code{\link{CheckDemogDB}},
  \code{\link{CheckMemory}},
  \code{\link{CheckRequirements}},
  \code{\link{CheckSourceYaml}},
  \code{\link{CheckSpaceDisk}}, \code{\link{ConnectToADL}},
  \code{\link{FormatList}}, \code{\link{GetConfiguration}},
  \code{\link{HttpCode}}, \code{\link{LoadDemogDB}},
  \code{\link{OptionsMessage}}, \code{\link{OrderColumn}},
  \code{\link{SaveLocal}}, \code{\link{TestNet}},
  \code{\link{TestVPN}}, \code{\link{TimeStamps}},
  \code{\link{Unzip}}, \code{\link{ValidateChunk}}
}
\author{
Jonathan Tooley Associados Lda
}
