% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetTimestampByObjInternal.R
\name{GetTimestampByObjInternal}
\alias{GetTimestampByObjInternal}
\title{GetTimestampByObjInternal}
\usage{
GetTimestampByObjInternal(Data, Project, ChunkName, FileExtension, Iterate)
}
\arguments{
\item{Data}{A TAP Chunk to be saved to the data lake}

\item{Project}{Name of the project. If the project doesn't exist a new folder will be created in the ADL.}

\item{ChunkName}{The name that is given to the data on the lake.}

\item{FileExtension}{A string containing either 'csv' or 'rdata' used to select the format in which the file is saved.}
}
\description{
Internal function that helps the GetTimeStampbyObj funtio to get the timestamp that already exists in the dataset
}
\seealso{
\code{\link{TAPChunks}}

Other Internal Utilities: \code{\link{AddDimension}},
  \code{\link{AddFirmographics}},
  \code{\link{AddHierarchy}}, \code{\link{CheckChunkInfo}},
  \code{\link{CheckDemogDB}}, \code{\link{CheckMemory}},
  \code{\link{CheckPackage}},
  \code{\link{CheckRequirements}},
  \code{\link{CheckSourceYaml}},
  \code{\link{CheckSpaceDisk}}, \code{\link{ConnectToADL}},
  \code{\link{CreateTable}}, \code{\link{DetectError}},
  \code{\link{FormatList}}, \code{\link{GetConfiguration}},
  \code{\link{GetMapping}},
  \code{\link{GetTimestampByObj}}, \code{\link{HttpCode}},
  \code{\link{LoadDemogDB}},
  \code{\link{OpenSqlConnection}},
  \code{\link{OptionsMessage}}, \code{\link{OrderColumn}},
  \code{\link{RMarkdown}},
  \code{\link{RecreateReadADLChunkExpression}},
  \code{\link{SaveLocal}}, \code{\link{ShowADLPath}},
  \code{\link{ShowADLPrefix}},
  \code{\link{ShowConfigurationPath}},
  \code{\link{TestNet}}, \code{\link{TestVPN}},
  \code{\link{TimeStamps}}, \code{\link{Unzip}},
  \code{\link{UpdateTAPChunks}},
  \code{\link{ValidateChunk}}, \code{\link{WriteFmt}}
}
\author{
JTA - The Data Scientists
}
\concept{Internal Utilities}
