% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GreenText.R
\name{GreenText}
\alias{GreenText}
\title{GreenText}
\usage{
GreenText(Text)
}
\arguments{
\item{Text}{String to apply font color}
}
\description{
Add Green font color
}
\examples{
  cat(GreenText("Hello World!"))
message(GreenText("Status100"))
}
\seealso{
Other Text Formatters: \code{\link{BlueText}},
  \code{\link{BoldText}}, \code{\link{RedText}},
  \code{\link{UnderLineText}}, \code{\link{YellowText}},
  \code{\link{catInfo}}
}
\keyword{internal}
