% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InternalUtilities.R
\name{HttpCode}
\alias{HttpCode}
\title{Decode the Http Error codes into messages}
\usage{
HttpCode(ErrorCode = 0)
}
\arguments{
\item{ErrorCode}{The client identifier}
}
\value{
Returns a message as a string
}
\description{
This function decode the Http Error codes into messages
}
\examples{
httpCode(200)
}
\seealso{
\code{\link{TAPChunks}}

Other Internal Utilities: \code{\link{AppendCSV}},
  \code{\link{CheckChunkInfo}}, \code{\link{CheckDemogDB}},
  \code{\link{CheckMemory}},
  \code{\link{CheckRequirements}},
  \code{\link{CheckSourceYaml}},
  \code{\link{CheckSpaceDisk}}, \code{\link{ConnectToADL}},
  \code{\link{FormatList}}, \code{\link{GetConfiguration}},
  \code{\link{GetMapping}}, \code{\link{LoadDemogDB}},
  \code{\link{OptionsMessage}}, \code{\link{OrderColumn}},
  \code{\link{SaveLocal}}, \code{\link{TestNet}},
  \code{\link{TestVPN}}, \code{\link{TimeStamps}},
  \code{\link{Unzip}}, \code{\link{ValidateChunk}}
}
\author{
Jonathan Tooley Associados Lda
}
\keyword{internal}
