% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ItalicText.R
\name{ItalicText}
\alias{ItalicText}
\title{Printing Italic Text}
\usage{
ItalicText(Text)
}
\arguments{
\item{Text}{String to apply Italic codes.}
}
\value{
Returns a string.
}
\description{
This function applies escape sequences to the sting that is passed to make
that string appear in italics.
}
\examples{
  cat(TAPChunks:::ItalicText("Hello World!"))
message(TAPChunks:::ItalicText("Status100"))
}
\seealso{
\code{\link{TAPChunks}}

Other Text Formatters: \code{\link{BlueText}},
  \code{\link{BoldText}}, \code{\link{GreenText}},
  \code{\link{RedText}}, \code{\link{StrikeText}},
  \code{\link{UnderLineText}}, \code{\link{YellowText}}
}
\author{
JTA - The Data Scientists
}
\concept{Text Formatters}
\keyword{internal}
