% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IterateScript.R
\name{IterateScript}
\alias{IterateScript}
\title{Run a script iteratively over a series of months}
\usage{
IterateScript(Script = "", From, To)
}
\arguments{
\item{Script}{The name of the script stored in your working directory or a full path name to where your
scrit is stored.  If a script name is not given the function will open an interactive dialog box for
the user to select their script.}

\item{From}{The start month from which we will run the script. This is a required field.}

\item{To}{The end month up until which we will run the script. This is a required field.}
}
\description{
Any system is limited by resources and telemetry data can be extremely large.
The TAP framework has a variety of solutions to allow for processing large quantities of
data and this is one of the most simple tools.  This function allows a user to take a TAP chunks
script that processes one month of data and then to run it repeatedly for many different months.
}
\section{Setting up an iteration}{

Your script may contain areas which will run over a series of months and also areas that are intended
to be run just once.  Typically a script has code to set up a function before iteration happens and code to
finish off
}

\seealso{
\code{\link{TAPChunks}}

Other Compute Tools: \code{\link{BeginIterate}},
  \code{\link{CleanMemory}}, \code{\link{EndIterate}},
  \code{\link{Protect}}
}
\author{
JTA: The Data Scientists
}
\concept{Compute Tools}
