% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ListCacheDirectory.R
\name{ListCacheDirectory}
\alias{ListCacheDirectory}
\title{List files in the local cache}
\usage{
ListCacheDirectory(Source = "", ShowAttributes = FALSE)
}
\arguments{
\item{Source}{Name of the source, if the Source parameter is not provided then a list of sources will be shown:\cr
\tabular{lll}{
\tab \strong{Source}       \tab \strong{Effect} \tab \strong{Data Documentation:}\cr
\tab \emph{Email}          \tab For the Spiceworks Email data \tab \code{\link{TestEmailChunk}}  \cr
\tab \emph{Client}         \tab For the Spiceworks PC Client data \tab \code{\link{TestClientChunk}} \cr
\tab \emph{Networkhv}      \tab For the Spiceworks Network data by hypervisor \tab \code{\link{TestNetworkhvChunk}} \cr
\tab \emph{Networkos}      \tab For the Spiceworks Network data by operating system \tab \code{\link{TestNetworkosChunk}} \cr
\tab \emph{ServerRoles}    \tab For the Spiceworks Server Roles data \tab \code{\link{TestServerRolesChunk}} \cr
\tab \emph{WorkloadDB}     \tab For the Spiceworks Workload data for database products \tab \code{\link{TestWorkloadDBChunk}} \cr
\tab \emph{WorkloadEco}    \tab This chunk lists the virt flag and data type along with Hypervisor and OS for the intersection of Network and VID \tab \code{\link{TestWorkloadEcoChunk}} \cr
\tab \emph{WorkloadHW}     \tab This chunk lists the metadata that we received relating to server hardware. \tab \code{\link{TestWorkloadHWChunk}} \cr
\tab \emph{WorkloadN}      \tab This chunk lists the virt flag and data type along with Hypervisor and OS for the Network style data \tab \code{\link{TestWorkloadNChunk}} \cr
\tab \emph{WorkloadVID}    \tab This chunk lists the virt flag and data type along with Hypervisor and OS for the VID style data \tab \code{\link{TestWorkloadVIDChunk}} \cr
\tab \emph{WorkloadWL}     \tab For the Spiceworks workload data for non-database products \tab \code{\link{TestWorkloadWLChunk}} \cr
\tab \emph{CCMCost}        \tab For the Spiceworks CCM Cost data\cr
\tab \emph{CCMUsage}       \tab For the Spiceworks CCM Usage data\cr
\tab \emph{OnlineServices} \tab For the Spiceworks Online Services Data\cr
\tab \emph{VID}            \tab For the Spiceworks VID Data\cr
\tab \emph{ServerAge}      \tab For the Spiceworks Server Age Data\cr
\tab \emph{Cosmos}         \tab For the Cosmos Data\cr
}}

\item{ShowAttributes}{When TRUE a detailed listing of the files attributes is shown.}
}
\value{
Returns a data table with a columns called 'File Name' holding strings with
the name of each file found.  If the user selected ShowAttributes = TRUE then the data
table will have additional columns for 'File Size' and 'File Created'
}
\description{
Show a list of files that we have in the local cache
}
\details{
This function will establish if the user has a local cache directory installed
in their machine.  If it doesn't exist it will be automatically created.
The function then returns a data table that has a list of chunks that are stored
in the cache.  \cr\cr
The user may select an optional filter using the source parameter
that will filter the results to show only chunks that relate to the source.\cr\cr
If the user selects showattributes = TRUE then the returned data table has
a series of properties listed that will help the user understand the formation of the
chunk. \cr
}
\examples{
ListCacheDirectory()
ListCacheDirectory(ShowAttributes = TRUE)
ListCacheDirectory("Client", ShowAttributes = TRUE)
}
\seealso{
\code{\link{TAPChunks}}

Other Cache Controls: \code{\link{CleanMemory}},
  \code{\link{DeleteCache}},
  \code{\link{DeleteOutdatedCache}}, \code{\link{Protect}},
  \code{\link{RefreshCache}}, \code{\link{ShowCachePath}},
  \code{\link{ShowCacheStatus}},
  \code{\link{ShowConfigurationPath}},
  \code{\link{Unprotect}}
}
\author{
JTA - The Data Scientists
}
