% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadDemogDB.R
\name{LoadDemogDB}
\alias{LoadDemogDB}
\title{Load the demographics database}
\usage{
LoadDemogDB()
}
\value{
A boolean status flag where TRUE shows the operation was successful
}
\description{
The TAP system uses a database of demographics to perform a variety of tasks:
 Filter Unclean orgs from datasets
 Compute distributions of firmographics to calculate weights
 Add firmographic data to other datasets to be used as slicers
 Define hierarchies in reports
 \cr
This function loads the database.
}
\details{
The database is stored in a separate environment called TAP_env.
The function tests that this environment exists and if it doesn't it will create the environment.
The function then tests for the presence of these three files:
 demog_helper  : This is a data table with all of the known Spiceworks organizations
 target_helper : This is a data table with the population targets that are provided by the Market Model
 unclean       : This is a vector of unclean organisation UUIDs
If they are all present then the routine end and returns a TRUE. If any of these files is missing
the system attempts to create the files by accessing from the Data Lake. If this is successful then it returns TRUE
}
\seealso{
\code{\link{TAPChunks}}

Other Internal Utilities: \code{\link{AddDimension}},
  \code{\link{AddFirmographics}},
  \code{\link{AddHierarchy}}, \code{\link{CheckChunkInfo}},
  \code{\link{CheckDemogDB}}, \code{\link{CheckMemory}},
  \code{\link{CheckPackage}},
  \code{\link{CheckRequirements}},
  \code{\link{CheckSourceYaml}},
  \code{\link{CheckSpaceDisk}}, \code{\link{ConnectToADL}},
  \code{\link{CreateTable}}, \code{\link{DetectError}},
  \code{\link{FormatList}}, \code{\link{GetConfiguration}},
  \code{\link{GetMapping}},
  \code{\link{GetTimestampByObjInternal}},
  \code{\link{GetTimestampByObj}}, \code{\link{HttpCode}},
  \code{\link{OpenSqlConnection}},
  \code{\link{OptionsMessage}}, \code{\link{OrderColumn}},
  \code{\link{RMarkdown}},
  \code{\link{RecreateReadADLChunkExpression}},
  \code{\link{SaveLocal}}, \code{\link{ShowADLPath}},
  \code{\link{ShowADLPrefix}},
  \code{\link{ShowConfigurationPath}},
  \code{\link{TestNet}}, \code{\link{TestVPN}},
  \code{\link{TimeStamps}}, \code{\link{Unzip}},
  \code{\link{UpdateTAPChunks}},
  \code{\link{ValidateChunk}}, \code{\link{WriteFmt}}
}
\author{
JTA - The Data Scientists
}
\concept{Internal Utilities}
\keyword{internal}
