% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InternalUtilities.R
\name{OpenSqlConnection}
\alias{OpenSqlConnection}
\title{Open SQL Connection}
\usage{
OpenSqlConnection(Server, Database, Uid = NULL, Pwd = NULL)
}
\arguments{
\item{Server}{name of server SQL}

\item{Database}{name of database SQL}

\item{Uid}{user identification of server SQL}

\item{Pwd}{password of server SQL}
}
\value{
Connection with SQL
}
\description{
This function can open connection with SQL without create a odbc connection
}
