% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InternalUtilities.R
\name{OrderColumn}
\alias{OrderColumn}
\title{Sort a data chunk's columns to a defined order}
\usage{
OrderColumn(DT)
}
\arguments{
\item{DT}{Chunk where the sort order will be applied}
}
\description{
Sort the columns in the data chunk
}
\seealso{
\code{\link{TAPChunks}}

Other Internal Utilities: \code{\link{AppendCSV}},
  \code{\link{CheckChunkInfo}}, \code{\link{CheckDemogDB}},
  \code{\link{CheckMemory}},
  \code{\link{CheckRequirements}},
  \code{\link{CheckSourceYaml}},
  \code{\link{CheckSpaceDisk}}, \code{\link{ConnectToADL}},
  \code{\link{FormatList}}, \code{\link{GetConfiguration}},
  \code{\link{GetMapping}}, \code{\link{HttpCode}},
  \code{\link{LoadDemogDB}}, \code{\link{OptionsMessage}},
  \code{\link{SaveLocal}}, \code{\link{TestNet}},
  \code{\link{TestVPN}}, \code{\link{TimeStamps}},
  \code{\link{Unzip}}, \code{\link{ValidateChunk}}
}
\author{
Jonathan Tooley Associados Lda
}
\keyword{internal}
