% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepareIPF.R
\name{PrepareIPF}
\alias{PrepareIPF}
\title{Preparing data for IPF}
\usage{
PrepareIPF(data = NULL, dimensions = NULL, metric = NULL, order = FALSE)
}
\arguments{
\item{data}{This can be a data table, a data frame or an array.  When the user presents a data table or data frame the system returns
an array in a format that is ready for the IPF function.  When the user presents an array the system returns a data table.}

\item{dimensions}{A vector of dimension names to include in the analysis}

\item{metric}{A calculation to be employed to generate the values which will be submitted to the analysis}
}
\value{
When the user submits a data frame or a data table the function returns an array.  When the user submits an array the system
returns a data table.
}
\description{
The mathematics of running an Iterative Proportional Fit are not hard
to do but the problem needs to be properly set up.  The data that is sent to IPF tools
such as \code{\link{RunIPF}} has many strict requirements and this function ensures that
data is properly formatted for IPF.
}
\section{Introduction}{

This function has the following objectives: \cr
\itemize{
\item It takes a table of data and analyses it for suitability for IPF. It checks
that there are columns of data that can form the dimensions of the IPF and it checks
for numerical data that can form the seed.
\item It allows the user to define both the dimensions and the metric that will form
the seed.  These definitions are validated and then applied.
\item It aggregates the data down to the chosen dimensionality so that we don't
get duplicate items in the seed.
\item It then checks that every coordinate in the dimension space is represented in
the data or fills blank coordinates.
\item It creates an n dimensional array of values that can be sent to the IPF function
\code{\link{RunIPF}}.
\item It reports all the above activity in a series of messages.
\item It can also take the n dimensional array format and reconvert this into a
data table.  This is used to return a completed IPF calculation back into a Chunk
format.
}
}

\section{Data Format}{

}

\section{Analysing Dimensions}{

}

\section{Defining Dimensions}{

}

\section{Analysing Metrics}{

}

\section{Defining Metrics}{

}

\section{Aggregating Data}{

}

\section{Filling Coordinates}{

}

\section{Creating an Array}{

}

\seealso{
\code{\link{TAPChunks}}

Other Iterative Proportional Fitting Tools: \code{\link{RunIPF}},
  \code{\link{SetTargets}}
}
\author{
JTA - The Data Scientists
}
