% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prints.R
\name{prints}
\alias{prints}
\alias{print.samplePackage_sampleFunction}
\title{S3Methods for Printing}
\usage{
\method{print}{sampleFunction}(x,  ...)
}
\arguments{
\item{x}{Object from \code{samplePackage} package}

\item{...}{Additional arguments}
}
\value{
Prints \code{samplePackage} object
}
\description{
Prints for \code{samplePackage} objects
}
