% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CacheControls.R
\name{Protect}
\alias{Protect}
\title{Protect a table in memory from deletion}
\usage{
Protect(...)
}
\arguments{
\item{...}{No parameters needed}
}
\value{
A object with attribute set as protect.
}
\description{
This function add an attribute that protects the object to be deleted
when the user uses the CleanMemory function.
}
\seealso{
\code{\link{TAPChunks}}

Other Cache Controls: \code{\link{CleanMemory}},
  \code{\link{DeleteCache}},
  \code{\link{DeleteOutdatedCache}},
  \code{\link{ListCacheDirectory}},
  \code{\link{RefreshCache}}, \code{\link{ShowCachePath}},
  \code{\link{ShowCacheStatus}},
  \code{\link{ShowConfigurationPath}},
  \code{\link{Unprotect}}
}
\author{
Jonathan Tooley Associados Lda
}
