% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataLake.R
\name{PublishADL}
\alias{PublishADL}
\title{Publish File into the Data Lake}
\usage{
PublishADL(Data, Project, ChunkName, FileExtension = "csv", Replace = F)
}
\arguments{
\item{Data}{This is a required field and is usually a TAP chunk but it can be any item of class "data.table" or "data.frame".
If the user submits a data frame then this will first be converted to a data table before the calculated column is added.
This is because data tables are far more efficient in R than a data frame.
The resulting table that is returned by this function will also be a data table even if a data frame was submitted.}

\item{Project}{Name of the project. If the project doesn't exist a new folder will be created in the ADL.}

\item{ChunkName}{Optional name to be given to the file when saved in the ADL. If null will aquire the object's name.}

\item{FileExtension}{Extension supports csv and Rdata files. By default is csv.}

\item{Replace}{Replace in case exists a file with same name. By default is FALSE.}
}
\description{
A function to save files in the Publication folder in the data lake. This function is similar to SaveADL() function, the difference is the chunk is saved in the Publish area.
By default saves a CSV file but also supports Rdata files.
}
\examples{
PublishADL(TestEmailChunk, Project="Emailtracker", ChunkName = "FactFile", FileExtension="csv", Replace = T)
}
\seealso{
\code{\link{TAPChunks}}

Other Data Lake: \code{\link{ListADLDirectory}},
  \code{\link{ReadABSFile}}, \code{\link{SaveADL}},
  \code{\link{UploadLocallyToADL}}
}
\author{
Jonathan Tooley Associados Lda
}
