% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InternalUtilities.R
\name{rbindImp}
\alias{rbindImp}
\title{Rbind function recreated to have a better perform in RAM memory}
\usage{
rbindImp(a, b)
}
\arguments{
\item{a}{first data table to append}

\item{b}{second data table to append}
}
\description{
This function recreates the base rbind from dta.table library. Keep in mind that it will remove the DT that you insert,
since it tries to reduce the memory usage.
}
\examples{
TODO
}
\seealso{
\code{\link{TAPChunks}}
}
\author{
Jonathan Tooley Associados Lda
}
