% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadABSFile.R
\name{ReadABSFile}
\alias{ReadABSFile}
\title{Download Files from Blob Storage}
\usage{
ReadABSFile(Source = "", Timestamp = "", FileSpecific = "",
  SavePath = "", Container = "spiceworks-tapdata")
}
\arguments{
\item{Source}{source to download}

\item{Timestamp}{month to download}

\item{FileSpecific}{Path to the file to be downloaded}

\item{SavePath}{TODO}

\item{Container}{Container of Blob Storage. By default is where spiceworks data is laying}
}
\value{
Prints a status message.
}
\description{
A function to download files from the Blob Storage in Azure Data Lake.
}
\details{
The raw tememtry data that we receive from SpiceWorks and other providers
is loaded to a Blob (\strong{B}inary \strong{L}arge \strong{Ob}ject) Store.
This data is then loaded via a managed process to our Azure data Lake in the \emph{chunk} format.
This function is used internally to process raw data from the Blob Store to the Data Lake.
It can, however, be used by anyone who wishes to access an original file.

The function is dependent of having the Azure Command Line tools installed on the host machine.

Be aware that the chunk format is much more efficient in terms of its use of both
disk and memory resources and is the preffered file format for telemetry analysis.

The requested file is downloaded as a csv to the users cache directory. It
is stored in a sub directory with the name of the month.

Any file may be downloaded by specifying its name in the \emph{FileSpecific}
parameter. If the user wishes to retrieve a telemetry file then there is a simpler
syntax where the function only requires the source and month to be specified
in the \emph{Source} and \emph{Timestamp} parameters.

If the user does not give parameters then the system will operate in an
interactive mode, offering a choice of options to the user.
}
\examples{
\donttest{ReadABSFile("Email", "2015M01")}
}
\seealso{
\code{\link{TAPChunks}}
#171

Other Data Lake Tools: \code{\link{ListADLDirectory}},
  \code{\link{ReadADLChunk}}, \code{\link{ReadADLFile}},
  \code{\link{UploadLocallyToADL}}
}
\author{
JTA - The Data Scientists
}
\concept{Data Lake Tools}
