% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecreateReadADLChunkExpression.R
\name{RecreateReadADLChunkExpression}
\alias{RecreateReadADLChunkExpression}
\title{Recreate ReadADLChunk expressions for the Iterate Wrapper}
\usage{
RecreateReadADLChunkExpression(Exp, Period)
}
\arguments{
\item{Exp}{This is an R expression that will have any periods replaced by the Period}

\item{Period}{This is a string with a standard TAP period in the format "2018M01" that
will be used to replace any existing periods passed in the expression.}
}
\value{
An object of class expression which, when run, should perform the same
as the expression send in the Exp parameter but for the Period defined in the Period parameter.
}
\description{
This is an internal function that is used by the iteration wrapper to
edit a user's script replacing any periods with the period that is required by the wrapper function.
The wrapper function is IterateScript().
}
\examples{
TAPChunks:::RecreateReadADLChunkExpression(Exp = parse(text = "Object <- ReadADLChunk(Source = 'Email', From = '2017M01')"), Period = "2018M02")
}
\seealso{
Other Internal Utilities: \code{\link{AddDimension}},
  \code{\link{AddFirmographics}},
  \code{\link{AddHierarchy}}, \code{\link{CheckChunkInfo}},
  \code{\link{CheckDemogDB}}, \code{\link{CheckMemory}},
  \code{\link{CheckPackage}},
  \code{\link{CheckRequirements}},
  \code{\link{CheckSourceYaml}},
  \code{\link{CheckSpaceDisk}}, \code{\link{ConnectToADL}},
  \code{\link{CreateTable}}, \code{\link{DetectError}},
  \code{\link{FormatList}}, \code{\link{GetConfiguration}},
  \code{\link{GetMapping}},
  \code{\link{GetTimestampByObjInternal}},
  \code{\link{GetTimestampByObj}}, \code{\link{HttpCode}},
  \code{\link{LoadDemogDB}},
  \code{\link{OpenSqlConnection}},
  \code{\link{OptionsMessage}}, \code{\link{OrderColumn}},
  \code{\link{RMarkdown}}, \code{\link{SaveLocal}},
  \code{\link{ShowADLPath}}, \code{\link{ShowADLPrefix}},
  \code{\link{ShowConfigurationPath}},
  \code{\link{TestNet}}, \code{\link{TestVPN}},
  \code{\link{TimeStamps}}, \code{\link{Unzip}},
  \code{\link{UpdateTAPChunks}},
  \code{\link{ValidateChunk}}, \code{\link{WriteFmt}}
}
\author{
JTA: The Data Scientists
}
\concept{Internal Utilities}
\keyword{internal}
