% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RedText.R
\name{RedText}
\alias{RedText}
\title{Printing Red Text}
\usage{
RedText(Text)
}
\arguments{
\item{Text}{String to add font color.}
}
\value{
Returns a string.
}
\description{
This function applies red font color to the text submitted by wrapping the text
with ANSI escape sequences.
}
\examples{
  cat(TAPChunks:::RedText("Hello World!"))
message(TAPChunks:::RedText("Status100"))
}
\seealso{
\code{\link{TAPChunks}}

Other Text Formatters: \code{\link{BlueText}},
  \code{\link{BoldText}}, \code{\link{GreenText}},
  \code{\link{ItalicText}}, \code{\link{StrikeText}},
  \code{\link{UnderLineText}}, \code{\link{YellowText}}
}
\author{
JTA - The Data Scientists
}
\concept{Text Formatters}
\keyword{internal}
