% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefreshCache.R
\name{RefreshCache}
\alias{RefreshCache}
\title{Detecting when data on the Data Lake has been updated}
\usage{
RefreshCache()
}
\value{
Returns an error status to show if the command completed succesfully or not
}
\description{
Delete and fetch superseded files
}
\details{
There are often times when our data provider reissues files after discovering
errors in the feed.  This then requires us to re process the loading of that data.
The issue is that, for the convenience of our users, we allow the local system to
keep a cached copy of the original chunk.  This command allows the user to detect
outdated files and to refresh the cache.
}
\examples{
RefreshCache()
}
\seealso{
\code{\link{TAPChunks}}

Other Cache Controls: \code{\link{CleanMemory}},
  \code{\link{DeleteCache}},
  \code{\link{DeleteOutdatedCache}},
  \code{\link{ListCacheDirectory}}, \code{\link{Protect}},
  \code{\link{ShowCachePath}},
  \code{\link{ShowCacheStatus}},
  \code{\link{ShowConfigurationPath}},
  \code{\link{Unprotect}}
}
\author{
JTA - The Data Scientists
}
