% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Manipulation.R
\name{RMarkdown}
\alias{RMarkdown}
\title{RMarkdown}
\usage{
RMarkdown(Source, MappingTable = source.os, Format, FileName)
}
\arguments{
\item{Source}{A data chunk}

\item{MappingTable}{The column name to map. If none, it has as default 'device_os' or 'os_name'}

\item{Format}{'Full' to show all OS mappings | 'Unknown' to show only OS that are unknown | 'NULL' to not create a report.}
}
\value{
Returns a html file that is saved in ShowCachePath()
}
\description{
Create a html file that is saved in ShowCachePath()
}
\details{
TODO
}
\seealso{
\code{\link{TAPChunks}}
Issue #

Other Chunk Manipulators: \code{\link{AddFirmographics}},
  \code{\link{AddOSHierarchy}},
  \code{\link{AggregateChunk}},
  \code{\link{CalculateColumn}}, \code{\link{FilterChunk}},
  \code{\link{JoinChunks}}, \code{\link{ShapeChunk}},
  \code{\link{UseAsSlicer}}
}
\author{
Jonathan Tooley Associados Lda
}
