% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunIPF.R
\name{RunIPF}
\alias{RunIPF}
\title{Run an Iterative Proportional Fit}
\usage{
RunIPF(M, target, maxiter = 25, tolerance = 1e-06)
}
\arguments{
\item{M}{A multidimensional array (or matrix) containing the seed for the IPF.}

\item{target}{A set of targets, one for each dimension, that the resulting array
will conform to.}

\item{maxiter}{The maximum number of iterations that the function will attempt
before stopping work.  If this is not specified the default is 25.}

\item{tolerance}{How close the function needs to get to the targets before it
declares that convergence has happened.}
}
\description{
Function to take an n-dimensional array and fit it to a set of
target vectors.
}
\seealso{
\code{\link{TAPChunks}}

Other Iterative Proportional Fitting Tools: \code{\link{PrepareIPF}},
  \code{\link{SetTargets}}
}
\author{
JTA - The Data Scientists
}
