% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleFunction.R
\name{sampleFunction}
\alias{sampleFunction}
\title{Sample function}
\usage{
sampleFunction(x)
}
\arguments{
\item{x}{The variable to use to compute the mean.}
}
\value{
The mean of x.
}
\description{
Sample function that provides the mean.
}
