% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaveADL.R
\name{SaveADL}
\alias{SaveADL}
\title{Save file in the Data Lake}
\usage{
SaveADL(Data, FileExtension = "csv", ADLDirectory, ChunkName, SavePath,
  SaveLocal = T)
}
\arguments{
\item{Data}{This is a required field and is usually a TAP chunk but it can be any item of class "data.table" or "data.frame".
If the user submits a data frame then this will first be converted to a data table before the calculated column is added.
This is because data tables are far more efficient in R than a data frame.
The resulting table that is returned by this function will also be a data table even if a data frame was submitted.}

\item{FileExtension}{Extension supports csv and Rdata files. By default is csv.}

\item{ADLDirectory}{Directory where the file will be stored. By default is a folder
with the user's alias.}

\item{ChunkName}{Optional name to be given to the file when saved in the ADL. If null will aquire the object's name.}

\item{SavePath}{Path to save the chunk locally. By default is TAP cache folder.}

\item{SaveLocal}{To save the chunk locally set as TRUE, otherwise set as FALSE. By default is TRUE.}
}
\value{
Returns a status message.
}
\description{
A function to save files in the data lake. By default saves a CSV file but also supports Rdata files.
By default saves in a folder named with user alias under user directory in the Data Lake.
}
\examples{
SaveADL(Data = TestEmailChunk, FileExtension = "csv", ADLDirectory = "Publication/", ChunkName = "EmailSample", SaveLocal=F)
}
\seealso{
\code{\link{TAPChunks}}

Other Reporting and saving tools: \code{\link{AppendCSV}},
  \code{\link{ApplyPenetration}},
  \code{\link{ExportToSQL}}, \code{\link{PublishADL}},
  \code{\link{UseAsSlicer}}
}
\author{
JTA - The Data Scientists
}
\concept{Reporting and saving tools}
