% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Manipulation.R
\name{SaveLocal}
\alias{SaveLocal}
\title{Save Objects}
\usage{
SaveLocal(..., FileExtension = "csv", Path = NULL)
}
\arguments{
\item{...}{Objects to be saved}

\item{FileExtension}{A string with the required file extension. This can be one of two types, csv or rdata.}

\item{Path}{Destination of the files with information of the R session objects}
}
\description{
This function allows to save R environment objects into local machine. Formats available csv and rdata
}
\examples{
SaveLocal(TestClientChunk, TestEmailChunk, path = getwd());
SaveLocal("TestClientChunk", "TestEmailChunk", FileExtension = "rdata", path = getwd())
}
\seealso{
\code{\link{TAPChunks}}

Other Internal Utilities: \code{\link{AppendCSV}},
  \code{\link{CheckChunkInfo}}, \code{\link{CheckDemogDB}},
  \code{\link{CheckMemory}},
  \code{\link{CheckRequirements}},
  \code{\link{CheckSourceYaml}},
  \code{\link{CheckSpaceDisk}}, \code{\link{ConnectToADL}},
  \code{\link{FormatList}}, \code{\link{GetConfiguration}},
  \code{\link{GetMapping}}, \code{\link{HttpCode}},
  \code{\link{LoadDemogDB}}, \code{\link{OptionsMessage}},
  \code{\link{OrderColumn}}, \code{\link{TestNet}},
  \code{\link{TestVPN}}, \code{\link{TimeStamps}},
  \code{\link{Unzip}}, \code{\link{ValidateChunk}}
}
\author{
Jonathan Tooley Associados Lda
}
