% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetTargets.R
\name{SetTargets}
\alias{SetTargets}
\title{Preparing a set of targets that will be used in a Iterative Proportional Fit algorithm}
\usage{
SetTargets(..., tolerance = 1e-06)
}
\arguments{
\item{...}{A list of named vectors (one for each dimension) with the set of targets}

\item{tolerance}{A value (defaults to 1E-6 if not provided) that will be used to weaken the test that each set of targets sums to the same value}
}
\value{
A list of named target vectors that will be used by the \emph{RunIPF} function.
This function will perform some extra checks to ensure that the dimension names given can be seen in the data
and that the number of values in each set matches the unique count of values in the dimension
as presented by the data.
}
\description{
A way to enter sets of targets to be used by the IPF function.
}
\section{Introduction}{

The function \code{\link{RunIPF}} requires some seed data and a series of target sets.  This function allows users to easily type target sets or reference
pre-calculated targets.
}

\section{Error Checking}{

The function performs some error checking on the values to make sure that the IPF will function well.  Tests include: \cr
\itemize{
\item At least two sets of targets must be submitted.
\item The target sets must all be named so that the IPF can link each target set to the correct dimension.
\item All the values in the target sets are numeric data types.
\item All the numeric values must be positive.
\item All of the target sets must sum to a series of value that are within tolerance of each other.
}
}

\section{Using the function}{

Using the function is very easy: Supply a list of vectors and give each one a name that matches your dimension. The
vectors can be referenced using the \emph{set} syntax \code{set(10,20,30,40)} and the name is
given using the = character to assign the set to the dimension.  The full syntax is then
\code{Segment = set(10,20,30,40)}. The complete command would then be:\cr
\code{SetTargets(Segment = set(10,20,30,40), Vertical = set(25,25,50)), ...} \cr
The function can accept vectors calculated elsewhere like this:\cr
\code{SetTargets(Segment = seg.vect, Vertical = vert.vect, ...)}\cr
}

\section{Ordering}{

When we give a set of numbers in each vector, these numbers will refer to the
desired total for a particular member of the dimension.  In order that the syntax
not get over complicated the particular member is not mentioned.  Instead the system
applies the numbers of the vector to the members of the dimension in a particular order.
Most dimensions in TAPChunks (such as geography and segment) have a specified order
that the system will use.  Data that has other dimensions without a predefined order
will be sorted into alphabetical order before the IPF is run.  \cr
The function \emph{PrepareIPF} will tell you the order in which the values of
each dimension will be used.  \cr
}

\examples{
SetTargets(geo = set(1, 2), vertical = set(1, 1, 0.999), segment = set(1, 1, 1.001), tolerance = 2E-3)

}
\seealso{
\code{\link{TAPChunks}}

Other Iterative Proportional Fitting Tools: \code{\link{PrepareIPF}},
  \code{\link{RunIPF}}
}
\author{
JTA - The Data Scientists
}
