% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShowCachePath.R
\name{ShowCachePath}
\alias{ShowCachePath}
\title{Defining where the cache folder is stored}
\usage{
ShowCachePath(Directory)
}
\arguments{
\item{Directory}{Optional Path to set the Cache location}
}
\value{
A string with the path to the cache folder.
}
\description{
This function allocates a path where the cache will be stored, if Directory parameter
is NULL then it will return the path where the cache has been established (usually in the user's
home directory).
}
\details{
When users run TAPChunks on their local machines it is normal for the system to establish the cache
in the user's home directory.  An example would be C:/Users/USERNAME/Documents/TAPCache.  If you call
this function without a parameter the system will return the path name of the cache directory.  When
you are using TAPChunks on a shared server you might need to change the location of your cache to
place it on a disk with more space. In this case you can establish a new location by passing it to the
function
}
\examples{
ShowCachePath()
\dontrun{ShowCachePath("F:")}
}
\seealso{
\code{\link{TAPChunks}}

Other Cache Controls: \code{\link{DeleteCache}},
  \code{\link{DeleteOutdatedCache}},
  \code{\link{ListCacheDirectory}},
  \code{\link{RefreshCache}},
  \code{\link{ShowCacheStatus}}, \code{\link{Unprotect}}
}
\author{
JTA - The Data Scientists
}
\concept{Cache Controls}
