% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShowConfigurationPath.R
\name{ShowConfigurationPath}
\alias{ShowConfigurationPath}
\title{Establishing and pointing to the configuration settings}
\usage{
ShowConfigurationPath(Directory)
}
\arguments{
\item{Directory}{When not given the function returns the path to the user's home
directory with \emph{TAPConfiguration} appended. If the user supplies a string
that refers to a directory that does not yet exist, a directory will be created.}
}
\value{
A string contining the configuration path
}
\description{
This is an internal function that will help to locate the directory
with configuration settings. The function is usually called without a parameter
and it returns the path to the configuration directory
(which is called TAPConfiguration and is stored in the user's home directory).
}
\examples{
TAPChunks:::ShowConfigurationPath()
}
\seealso{
\code{\link{TAPChunks}}

Other Internal Utilities: \code{\link{AddDimension}},
  \code{\link{AddFirmographics}},
  \code{\link{AddHierarchy}}, \code{\link{CheckChunkInfo}},
  \code{\link{CheckDemogDB}}, \code{\link{CheckMemory}},
  \code{\link{CheckPackage}},
  \code{\link{CheckRequirements}},
  \code{\link{CheckSourceYaml}},
  \code{\link{CheckSpaceDisk}}, \code{\link{ConnectToADL}},
  \code{\link{CreateTable}}, \code{\link{DetectError}},
  \code{\link{FormatList}}, \code{\link{GetConfiguration}},
  \code{\link{GetMapping}},
  \code{\link{GetTimestampByObjInternal}},
  \code{\link{GetTimestampByObj}}, \code{\link{HttpCode}},
  \code{\link{LoadDemogDB}},
  \code{\link{OpenSqlConnection}},
  \code{\link{OptionsMessage}}, \code{\link{OrderColumn}},
  \code{\link{RMarkdown}},
  \code{\link{RecreateReadADLChunkExpression}},
  \code{\link{SaveLocal}}, \code{\link{ShowADLPath}},
  \code{\link{ShowADLPrefix}}, \code{\link{TestNet}},
  \code{\link{TestVPN}}, \code{\link{TimeStamps}},
  \code{\link{Unzip}}, \code{\link{UpdateTAPChunks}},
  \code{\link{ValidateChunk}}, \code{\link{WriteFmt}}
}
\author{
JTA - The Data Scientists
}
\concept{Internal Utilities}
\keyword{internal}
