% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Manipulation.R
\name{ShowLongitudinalSize}
\alias{ShowLongitudinalSize}
\title{Show the Size of a Longitudinal Set}
\usage{
ShowLongitudinalSize(Data)
}
\arguments{
\item{Data}{TODO}
}
\examples{
ShowLongitudinalSize(TestEmailChunk)
}
\author{
Jonathan Tooley Associated Lda
}
