% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Manipulation.R
\name{ShowSampleSize}
\alias{ShowSampleSize}
\title{Summarizing a file to see the sample size}
\usage{
ShowSampleSize(Data, By = NULL)
}
\arguments{
\item{Data}{TODO}

\item{By}{TODO}
}
\description{
This function will tell the user the sample size in a file.
The sample size is defined as the number of distinct orgs in the data and is
split between the number of unclean orgs, unweighted clean orgs and weighted clean
orgs.
}
\examples{
ShowSampleSize(TestEmailChunk)
ShowSampleSize(ApplyWeightCalc(TestEmailChunk))
ShowSampleSize(ApplyWeightCalc(TestEmailChunk), By = "email_delivery")
}
\author{
JTA Associados Lda
}
