% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Document.R
\name{static_help}
\alias{static_help}
\title{Preparing Static HTML Documentation for an R Package}
\usage{
static_help(pkg, links = tools::findHTMLlinks())
}
\arguments{
\item{pkg}{Name of the package to be documented}

\item{links}{A vector with the HTML links to be used.  Defaults to findHTMLlinks.}
}
\description{
A function to write the R manual content into a Hyperlinked HTML format.
}
\details{
When R builds a package it writes an index file as HTML in the package
directory and bundles the help text into a specific format.  This function unpacks
the help text and creates a set of HTML files, one for each manual entry. The HTML
is hyperlinked.
}
\examples{
\donttest{TAPChunks:::static_help("TAPChunks")}
}
\seealso{
Other Documentation Support Tools: \code{\link{doc_cols}},
  \code{\link{doc_content}}, \code{\link{doc_variables}}
}
\author{
JTA Associados Lda
}
\keyword{internal}
