% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StrikeText.R
\name{StrikeText}
\alias{StrikeText}
\title{Printing Struck Text}
\usage{
StrikeText(Text)
}
\arguments{
\item{Text}{String to apply Strike codes.}
}
\value{
Returns a string.
}
\description{
This function uses escape sequences to add codes to the string that
is passed to this function so that string appears to be struck out.
}
\examples{
cat(TAPChunks:::StrikeText("Hello World!"))
message(TAPChunks:::StrikeText("Status100"))
}
\seealso{
\code{\link{TAPChunks}}

Other Text Formatters: \code{\link{BlueText}},
  \code{\link{BoldText}}, \code{\link{GreenText}},
  \code{\link{ItalicText}}, \code{\link{RedText}},
  \code{\link{UnderLineText}}, \code{\link{YellowText}}
}
\author{
JTA - The Data Scientists
}
\concept{Text Formatters}
\keyword{internal}
