% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestClientChunk.R
\docType{data}
\name{TestClientChunk}
\alias{TestClientChunk}
\alias{Client}
\alias{client}
\alias{apple_icloud}
\alias{box}
\alias{cisco_webex}
\alias{dropbox}
\alias{google_apps}
\alias{google_chrome}
\alias{google_drive}
\alias{internet_explorer}
\alias{itunes}
\alias{mozilla_firefox}
\alias{msft_office_2000}
\alias{msft_office_2003_basic}
\alias{msft_office_2003_pro}
\alias{msft_office_2007_basic}
\alias{msft_office_2007_pro_plus}
\alias{msft_office_2007_smb}
\alias{msft_office_2010_home}
\alias{msft_office_2010_pro}
\alias{msft_office_2010_pro_plus}
\alias{msft_office_2010_standard_revised}
\alias{msft_office_2013_pro}
\alias{msft_office_2013_pro_plus}
\alias{msft_office_2013_standard}
\alias{msft_office_365}
\alias{msft_skydrive}
\alias{msft_works}
\alias{pandora}
\alias{skype}
\title{Test Client data}
\description{
This is an example of a raw Client chunk as we receive them from SpiceWorks. The
data file is limited to 500 records and relates to just one period - 2016M01.
}
\section{File Columns}{

\tabular{ll}{
\strong{Column Name} \tab \strong{Description} \cr
\emph{row_id} \tab This is the identifier of the original row from where the data was obtained.
This allows us to recreate the exact same representation that we received from SpiceWorks. \cr
\emph{id} \tab This is the original id applied by Spiceworks. Its use is deprecated.\cr
\emph{uuid} \tab This is the standard org identifier for Spiceworks files.\cr
\emph{msft_office_2010_standard_revised_flag} \tab \cr
\emph{iwork_ofm_ie_flag} \tab \cr
\emph{slack_flag} \tab \cr
\emph{oem} \tab \cr
\emph{form_factor} \tab \cr
\emph{os_name} \tab \cr
\emph{timestamp} \tab The month that the data was delivered from Spiceworks in the format 2015M01.\cr
\emph{variable} \tab The name of the original Spiceworks product column.\cr
\emph{value}\tab The value that was in the original column.\cr}
}

\section{Example Content}{

\tabular{rlrrrlllllr}{
\strong{id} \tab \strong{uuid} \tab \strong{msft_office_2010_standard_revised_flag} \tab \strong{iwork_ofm_ie_flag} \tab \strong{slack_flag} \tab \strong{oem} \tab \strong{form_factor} \tab \strong{os_name} \tab \strong{timestamp} \tab \strong{variable} \tab \strong{value}\cr
903 \tab f50de... \tab 1 \tab 1 \tab 1 \tab asustek  \tab L \tab Windows 10 Pro \tab 2016M01 \tab device_count              \tab 1\cr
904 \tab f50de... \tab 1 \tab 1 \tab 1 \tab gigabyte \tab D \tab Windows 10 Pro \tab 2016M01 \tab device_count              \tab 1\cr
905 \tab f50de... \tab 1 \tab 1 \tab 1 \tab lenovo   \tab L \tab Windows 10 Pro \tab 2016M01 \tab device_count              \tab 1\cr
906 \tab f50de... \tab 1 \tab 1 \tab 1 \tab gigabyte \tab D \tab Windows 10 Pro \tab 2016M01 \tab device_count              \tab 1\cr
907 \tab f50de... \tab 1 \tab 1 \tab 1 \tab Other    \tab D \tab Windows 10 Pro \tab 2016M01 \tab device_count              \tab 1\cr
904 \tab f50de... \tab 1 \tab 1 \tab 1 \tab gigabyte \tab D \tab Windows 10 Pro \tab 2016M01 \tab msft_office_2013_pro_plus \tab 1\cr
906 \tab f50de... \tab 1 \tab 1 \tab 1 \tab gigabyte \tab D \tab Windows 10 Pro \tab 2016M01 \tab msft_office_2013_pro_plus \tab 1\cr
907 \tab f50de... \tab 1 \tab 1 \tab 1 \tab Other    \tab D \tab Windows 10 Pro \tab 2016M01 \tab msft_office_2013_pro_plus \tab 1\cr
907 \tab f50de... \tab 1 \tab 1 \tab 1 \tab Other    \tab D \tab Windows 10 Pro \tab 2016M01 \tab google_drive              \tab 1
}
}

\section{Variables Included}{

\tabular{ll}{
\strong{variable} \tab \strong{Notes}\cr
\emph{ apple_icloud } \tab \cr
\emph{ box } \tab \cr
\emph{ cisco_webex } \tab \cr
\emph{ device_count } \tab \cr
\emph{ dropbox } \tab \cr
\emph{ google_apps } \tab \cr
\emph{ google_chrome } \tab \cr
\emph{ google_drive } \tab \cr
\emph{ internet_explorer } \tab \cr
\emph{ itunes } \tab \cr
\emph{ mozilla_firefox } \tab \cr
\emph{ msft_office_2000 } \tab \cr
\emph{ msft_office_2003_basic } \tab \cr
\emph{ msft_office_2003_pro } \tab \cr
\emph{ msft_office_2007_basic } \tab \cr
\emph{ msft_office_2007_pro_plus } \tab \cr
\emph{ msft_office_2007_smb } \tab \cr
\emph{ msft_office_2010_home } \tab \cr
\emph{ msft_office_2010_pro } \tab \cr
\emph{ msft_office_2010_pro_plus } \tab \cr
\emph{ msft_office_2010_standard_revised } \tab \cr
\emph{ msft_office_2013_pro } \tab \cr
\emph{ msft_office_2013_pro_plus } \tab \cr
\emph{ msft_office_2013_standard } \tab \cr
\emph{ msft_office_365 } \tab \cr
\emph{ msft_skydrive } \tab \cr
\emph{ msft_works } \tab \cr
\emph{ pandora } \tab \cr
\emph{ skype }\tab \cr}
}

\seealso{
\code{\link{TAPChunks}}

Other Test Data: \code{\link{TestCCMCostChunk}},
  \code{\link{TestCCMUsageChunk}},
  \code{\link{TestEmailChunk}},
  \code{\link{TestNetworkhvChunk}},
  \code{\link{TestNetworkosChunk}},
  \code{\link{TestOnlineServicesChunk}},
  \code{\link{TestServerRolesChunk}},
  \code{\link{TestVIDChunk}},
  \code{\link{TestWorkloadDBChunk}},
  \code{\link{TestWorkloadEcoChunk}},
  \code{\link{TestWorkloadHWChunk}},
  \code{\link{TestWorkloadNChunk}},
  \code{\link{TestWorkloadVIDChunk}},
  \code{\link{TestWorkloadWLChunk}}
}
\author{
JTA - The Data Scientists
}
\keyword{internal}
