% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TAPChunks.R
\docType{data}
\name{TestServerRolesChunk}
\alias{TestServerRolesChunk}
\alias{ad_cert_svcs_16}
\alias{ad_domain_svcs_10}
\alias{ad_fed_svcs_8}
\alias{ad_lite_dir_svcs_9}
\alias{apps_1}
\alias{branchcache_324}
\alias{cat_base}
\alias{cat_born_cloud}
\alias{cat_cloud_infra}
\alias{cat_legacy}
\alias{dhcp_12}
\alias{dns_13}
\alias{dotnet_351_36}
\alias{dotnet_45_features_466}
\alias{essntls_exprnce_485}
\alias{failover_cluster_33}
\alias{file_6}
\alias{file_and_storage_481}
\alias{group_policy_69}
\alias{hyperv_20}
\alias{msg_queuing_49}
\alias{multipath_io_57}
\alias{powershell_417}
\alias{powershell_int_scripting_351}
\alias{print_7}
\alias{remote_admin_67}
\alias{remote_desktop_18}
\alias{telnet_client_44}
\alias{web_2}
\alias{win_activation_svcs_41}
\alias{win_deploy_svcs_19}
\alias{win_powershell_66}
\alias{win_server_update_21}
\alias{wins_svr_40}
\alias{winsvr_backup_39}
\alias{wow64_support_340}
\alias{ServerRoles}
\alias{serverroles}
\title{Test ServerRoles data}
\description{
This is an example of a raw Server Role chunk as we receive them from SpiceWorks. The
data file is limited to 500 records and relates to just one period - 2016M01.
}
\section{File Columns}{

\tabular{ll}{
\strong{Column Name} \tab \strong{Description} \cr
\emph{row_id} \tab This is the identifier of the original row from where the data was obtained.
This allows us to recreate the exact same representation that we received from SpiceWorks.\cr
\emph{id} \tab \cr
\emph{uuid} \tab This is the standard org identifier for Spiceworks files.\cr
\emph{server_roles_uuid} \tab \cr
\emph{timestamp} \tab \cr
\emph{variable} The name of the original Spiceworks Product Column.\tab \cr
\emph{value}\tab \cr}
}

\section{Example Content}{

\tabular{rllllr}{
\strong{id} \tab \strong{uuid} \tab \strong{server_roles_uuid} \tab \strong{timestamp} \tab \strong{variable} \tab \strong{value}\cr
1 \tab 0235a... \tab 2fa78c8.... \tab 2016M01 \tab cat_base               \tab 1\cr
1 \tab 0235a... \tab 2fa78c8.... \tab 2016M01 \tab file_and_storage_481   \tab 1\cr
1 \tab 0235a... \tab 2fa78c8.... \tab 2016M01 \tab cat_born_cloud         \tab 1\cr
1 \tab 0235a... \tab 2fa78c8.... \tab 2016M01 \tab web_2                  \tab 1\cr
1 \tab 0235a... \tab 2fa78c8.... \tab 2016M01 \tab dotnet_45_features_466 \tab 1\cr
1 \tab 0235a... \tab 2fa78c8.... \tab 2016M01 \tab cat_cloud_infra        \tab 1\cr
1 \tab 0235a... \tab 2fa78c8.... \tab 2016M01 \tab powershell_417         \tab 1\cr
1 \tab 0235a... \tab 2fa78c8.... \tab 2016M01 \tab cat_legacy             \tab 1\cr
1 \tab 0235a... \tab 2fa78c8.... \tab 2016M01 \tab wow64_support_340      \tab 1
}
}

\section{Variables Included}{

\tabular{ll}{
\strong{variable} \tab \strong{Notes}\cr
\emph{ ad_cert_svcs_16 } \tab \cr
\emph{ ad_domain_svcs_10 } \tab \cr
\emph{ ad_fed_svcs_8 } \tab \cr
\emph{ ad_lite_dir_svcs_9 } \tab \cr
\emph{ apps_1 } \tab \cr
\emph{ branchcache_324 } \tab \cr
\emph{ cat_base } \tab \cr
\emph{ cat_born_cloud } \tab \cr
\emph{ cat_cloud_infra } \tab \cr
\emph{ cat_legacy } \tab \cr
\emph{ dhcp_12 } \tab \cr
\emph{ dns_13 } \tab \cr
\emph{ dotnet_351_36 } \tab \cr
\emph{ dotnet_45_features_466 } \tab \cr
\emph{ essntls_exprnce_485 } \tab \cr
\emph{ failover_cluster_33 } \tab \cr
\emph{ file_6 } \tab \cr
\emph{ file_and_storage_481 } \tab \cr
\emph{ group_policy_69 } \tab \cr
\emph{ hyperv_20 } \tab \cr
\emph{ msg_queuing_49 } \tab \cr
\emph{ multipath_io_57 } \tab \cr
\emph{ powershell_417 } \tab \cr
\emph{ powershell_int_scripting_351 } \tab \cr
\emph{ print_7 } \tab \cr
\emph{ remote_admin_67 } \tab \cr
\emph{ remote_desktop_18 } \tab \cr
\emph{ telnet_client_44 } \tab \cr
\emph{ web_2 } \tab \cr
\emph{ win_activation_svcs_41 } \tab \cr
\emph{ win_deploy_svcs_19 } \tab \cr
\emph{ win_powershell_66 } \tab \cr
\emph{ win_server_update_21 } \tab \cr
\emph{ wins_svr_40 } \tab \cr
\emph{ winsvr_backup_39 } \tab \cr
\emph{ wow64_support_340 }\tab \cr}
}

\seealso{
\code{\link{TAPChunks}}

Other Test Data: \code{\link{TestCCMCostChunk}},
  \code{\link{TestCCMUsageChunk}},
  \code{\link{TestClientChunk}},
  \code{\link{TestEmailChunk}},
  \code{\link{TestNetworkhvChunk}},
  \code{\link{TestNetworkosChunk}},
  \code{\link{TestOnlineServicesChunk}},
  \code{\link{TestVIDChunk}},
  \code{\link{TestWorkloadDBChunk}},
  \code{\link{TestWorkloadEcoChunk}},
  \code{\link{TestWorkloadHWChunk}},
  \code{\link{TestWorkloadNChunk}},
  \code{\link{TestWorkloadVIDChunk}},
  \code{\link{TestWorkloadWLChunk}}
}
\author{
Jonathan Tooley Associados Lda
}
\keyword{internal}
