% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InternalUtilities.R
\name{TimeStamps}
\alias{TimeStamps}
\title{Preparing a range of timestamps}
\usage{
TimeStamps(From = "", To = "")
}
\arguments{
\item{From}{A timestamp to indicate the beginning of the range}

\item{To}{A timestamp to indicate the end of the range}
}
\value{
Returns a vector of all posible timestamps between from and to
}
\description{
A function that will accept the start and end timestamps and turn this information
into a vector with has all of the timestamps from the first until the last exclusively.
}
\details{
This function may be used to conserve memory when dealing with large files.
The function returns a vector with the range of timestamps so that a script may process
the chunks on a month by month basis, clearing the memory between each month.
}
\examples{
TimeStamps(from = "2017M01", to = "2017M12")
}
\seealso{
\code{\link{TAPChunks}}

Other Internal Utilities: \code{\link{AppendCSV}},
  \code{\link{CheckChunkInfo}}, \code{\link{CheckDemogDB}},
  \code{\link{CheckMemory}},
  \code{\link{CheckRequirements}},
  \code{\link{CheckSourceYaml}},
  \code{\link{CheckSpaceDisk}}, \code{\link{ConnectToADL}},
  \code{\link{FormatList}}, \code{\link{GetConfiguration}},
  \code{\link{GetMapping}}, \code{\link{HttpCode}},
  \code{\link{LoadDemogDB}}, \code{\link{OptionsMessage}},
  \code{\link{OrderColumn}}, \code{\link{SaveLocal}},
  \code{\link{TestNet}}, \code{\link{TestVPN}},
  \code{\link{Unzip}}, \code{\link{ValidateChunk}}
}
\author{
Jonathan Tooley Associados Lda
}
