% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnderLineText.R
\name{UnderLineText}
\alias{UnderLineText}
\title{Printing UnderLined Text}
\usage{
UnderLineText(Text = "")
}
\arguments{
\item{Text}{String to apply underline codes.}
}
\value{
Returns a string.
}
\description{
This function underlines the text submitted by wrapping the text
with ANSI escape sequences.
}
\examples{
  cat(TAPChunks:::UnderLineText("Hello World!"))
message(TAPChunks:::UnderLineText("Status 100"))
}
\seealso{
\code{\link{TAPChunks}}

Other Text Formatters: \code{\link{BlueText}},
  \code{\link{BoldText}}, \code{\link{GreenText}},
  \code{\link{ItalicText}}, \code{\link{RedText}},
  \code{\link{StrikeText}}, \code{\link{YellowText}}
}
\author{
JTA - The Data Scientists
}
\concept{Text Formatters}
\keyword{internal}
