% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CacheControls.R
\name{Unprotect}
\alias{Unprotect}
\title{Unprotect the object that has been assigned as Protected}
\usage{
Unprotect(...)
}
\arguments{
\item{...}{No parameters needed}
}
\value{
A object without attribute set as protect.
}
\description{
This function removes the Protect attribute from the object.
When the user uses the CleanMemory function the object will be deleted.
}
\seealso{
\code{\link{TAPChunks}}

Other Cache Controls: \code{\link{CleanMemory}},
  \code{\link{DeleteCache}},
  \code{\link{DeleteOutdatedCache}},
  \code{\link{ListCacheDirectory}}, \code{\link{Protect}},
  \code{\link{RefreshCache}}, \code{\link{ShowCachePath}},
  \code{\link{ShowCacheStatus}},
  \code{\link{ShowConfigurationPath}}
}
\author{
Jonathan Tooley Associados Lda
}
