% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InternalUtilities.R
\name{Unzip}
\alias{Unzip}
\title{Unzip files}
\usage{
Unzip(File = "", SavePath)
}
\arguments{
\item{File}{TODO}

\item{Savepath}{TODO}
}
\description{
TODO
}
\seealso{
\code{\link{TAPChunks}}

Other Internal Utilities: \code{\link{AppendCSV}},
  \code{\link{CheckChunkInfo}}, \code{\link{CheckDemogDB}},
  \code{\link{CheckMemory}},
  \code{\link{CheckRequirements}},
  \code{\link{CheckSourceYaml}},
  \code{\link{CheckSpaceDisk}}, \code{\link{ConnectToADL}},
  \code{\link{FormatList}}, \code{\link{GetConfiguration}},
  \code{\link{GetMapping}}, \code{\link{HttpCode}},
  \code{\link{LoadDemogDB}}, \code{\link{OptionsMessage}},
  \code{\link{OrderColumn}}, \code{\link{SaveLocal}},
  \code{\link{TestNet}}, \code{\link{TestVPN}},
  \code{\link{TimeStamps}}, \code{\link{ValidateChunk}}
}
\author{
Jonathan Tooley Associados Lda
}
\keyword{internal}
