% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Configuration.R
\name{UpdateTAPChunks}
\alias{UpdateTAPChunks}
\title{UpdateTAPChunks}
\usage{
UpdateTAPChunks()
}
\value{
None
}
\description{
Update TAPChunks version in case there is a more recent one.
}
\examples{
UpdatePackage()
}
\seealso{
\code{\link{TAPChunks}}
}
\author{
Jonathan Tooley Associados Lda
}
