% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataLake.R
\name{UploadLocallyToADL}
\alias{UploadLocallyToADL}
\title{Save file in the Data Lake}
\usage{
UploadLocallyToADL(FileName, FileExtension = "", ADLDirectory,
  FileDirectory = "")
}
\arguments{
\item{FileName}{Name file}

\item{FileExtension}{Extension file}

\item{ADLDirectory}{Directory where the file will be stored.}

\item{FileDirectory}{TODO}
}
\value{
Prints a status message.
}
\description{
A function to save files in the data lake.
}
\examples{
TODO
}
\seealso{
\code{\link{TAPChunks}}

Other Data Lake: \code{\link{ListADLDirectory}},
  \code{\link{PublishADL}}, \code{\link{ReadABSFile}},
  \code{\link{SaveADL}}
}
\author{
Jonathan Tooley Associados Lda
}
