% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Manipulation.R
\name{UseAsSlicer}
\alias{UseAsSlicer}
\title{UseAsSlicer}
\usage{
UseAsSlicer(Data, ReportName = "test")
}
\arguments{
\item{Data}{A data table}

\item{ReportName}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\details{
TODO
}
\seealso{
\code{\link{TAPChunks}}
Issue #129

Other Chunk Manipulators: \code{\link{AddFirmographics}},
  \code{\link{AddOSHierarchy}},
  \code{\link{AggregateChunk}},
  \code{\link{CalculateColumn}}, \code{\link{FilterChunk}},
  \code{\link{JoinChunks}}, \code{\link{RMarkdown}},
  \code{\link{ShapeChunk}}
}
\author{
Jonathan Tooley Associados Lda
}
