% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Manipulation.R
\name{WeightCalculation}
\alias{WeightCalculation}
\title{Reviewing a weight calculation}
\usage{
WeightCalculation(Data, Granularity = "gsv")
}
\details{
DEPRECATED: See \code{\link{ShowWeightCalc}}
}
\keyword{internal}
