% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wiki_hot_100s}
\alias{wiki_hot_100s}
\title{Data for songs}
\format{
A data frame with 5701 rows and 4 variables:
\describe{
 \item{no}{Number of the song}
 \item{title}{Title of the song}
 \item{artisit}{Artist name}
 \item{year}{Release year}
}
}
\source{
{Exported from R package "billboard"}
}
\usage{
wiki_hot_100s
}
\description{
Contains songs information
}
\examples{
 data(wiki_hot_100s)
}
\keyword{datasets}
