% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InternalUtilities.R
\name{WriteFmt}
\alias{WriteFmt}
\title{WriteFmt (Internal Function used in ExportToSQL)}
\usage{
WriteFmt(InputData, ServerPath = ".")
}
\arguments{
\item{InputData}{TODO}

\item{ServerPath}{TODO}
}
\description{
TODO
}
\examples{
TODO
}
\seealso{
\code{\link{TAPChunks}}
}
\author{
Jonathan Tooley Associados Lda
}
