% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Xcel.R
\name{Xcel}
\alias{Xcel}
\title{Pass Data to an Excel file}
\usage{
Xcel(Data)
}
\arguments{
\item{Data}{Dataset to be opened in Excel}
}
\description{
This function takes the data chunk that is passed and
saves its content in \emph{Tab Separated Values format} in a temporary file stored in
the Cache. The system then asks the operating system to open the file in Excel. If you add
any logic to your excel you will need to save your work in .xlsx format.
Excel must be installed on your system.
}
\examples{
Xcel(TestEmailChunk)
}
\seealso{
\code{\link{TAPChunks}}
}
\author{
JTA - The Data Scientists
}
\concept{Utilities}
