% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YellowText.R
\name{YellowText}
\alias{YellowText}
\title{YellowText}
\usage{
YellowText(Text)
}
\arguments{
\item{Text}{String to apply font color}
}
\description{
Apply yellow font color
}
\examples{
  cat(YellowText("Hello World!"))
message(YellowText("Status100"))
}
\seealso{
Other Text Formatters: \code{\link{BlueText}},
  \code{\link{BoldText}}, \code{\link{GreenText}},
  \code{\link{RedText}}, \code{\link{UnderLineText}},
  \code{\link{catInfo}}
}
\keyword{internal}
