% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Manipulation.R
\name{AddDimension}
\alias{AddDimension}
\title{ADD Dimension}
\usage{
AddDimension(Data, Dimension = "G", SelectHierarchy = NULL)
}
\arguments{
\item{Data}{TODO}

\item{Dimension}{TODO}

\item{SelectHierarchy}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\details{
TODO
}
\examples{
AddDimension(Data = Data, Dimension = "P")
}
\author{
JTA Associados Lda
}
