% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Manipulation.R
\name{AddHierarchy}
\alias{AddHierarchy}
\title{Add Hierarchy}
\usage{
AddHierarchy(Data, Dimension = NULL, Hierarchy = NULL)
}
\arguments{
\item{Data}{A data chunk}

\item{Dimension}{Dimension to be added to the dataset. i.e "G" "S" "O" "P"}

\item{Hierarchy}{TODO}
}
\description{
This function allows adding all hierarchies at same time. The user can define which dimension wants to add to the dataset.
}
\section{Reference Data Repository}{

The Reference Data Repository (RDR) allows stores master data relating to
our taxonomy.  It can be used to add taxonomy hierarchies to chunks that have had
firmographic data added (See \code{\link{AddFirmographics}}). Firmographic data is
established by our data load processes (which refer to the RDR to allocate firmographics)
and is stored in a separate database.

Typically functions using the RDR will try and use firmographic data in the chunk
submitted but if they can't detect firmographic data they will attempt to add this by calling
AddFirmographics for you.

Functions that access the RDR require an internet connection which has access to
MS Corp net.
}

\section{Geo Hierarchy}{

The function accesses the RDR and retrieves the standard geography
hierarchy that is matched to the country names that the file has.
}

\section{Segment Hierarchy}{

The function accesses the RDR and retrieves the standard geography
hierarchy that is matched to the segment names that the file has.
}

\section{Product Hierarchy}{

The raw data that we receive from SpiceWorks contains a great deal of product
information.  This is typically shown in the many product columns that we see in a typical
raw file and that gets transposed when written as a chunk.
The product descriptions are detailed and so to summarize our finindings it is necessary
to add a product hierarchy.
}

\seealso{
\code{\link{TAPChunks}}

Other RDR Integration Tools: \code{\link{AddGeoHierarchy}},
  \code{\link{AddOSHierarchy}},
  \code{\link{AddProductHierarchy}},
  \code{\link{AddSegmentHierarchy}},
  \code{\link{AddVerticalHierarchy}}
}
\author{
Jonathan Tooley Associados Lda
}
