% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Manipulation.R
\name{AggregateChunk}
\alias{AggregateChunk}
\title{Aggregate chunks to create smaller chunks}
\usage{
AggregateChunk(Data, Filter = 1 == 1, Aggregation = .N, By = NULL, ...)
}
\arguments{
\item{Data}{Chunk to be aggregated}

\item{Filter}{Filter to be applyied in the chunk}

\item{Aggregation}{Aggregate function and column to be summarized.}

\item{By}{columns where the chunk will be grouped by.}

\item{...}{Error handelling}
}
\description{
Aggregate chunks into a smaller chunks by the metadata selected by the user. (Review)
}
\seealso{
\code{\link{TAPChunks}}

Other Chunk Manipulators: \code{\link{AddFirmographics}},
  \code{\link{AddOSHierarchy}},
  \code{\link{CalculateColumn}}, \code{\link{FilterChunk}},
  \code{\link{JoinChunks}}, \code{\link{RMarkdown}},
  \code{\link{ShapeChunk}}, \code{\link{UseAsSlicer}}
}
\author{
Jonathan Tooley Associados Lda
}
