% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Manipulation.R
\name{AppendCSV}
\alias{AppendCSV}
\title{Append CSV}
\usage{
AppendCSV(..., Folder, Name, Remove = F)
}
\arguments{
\item{...}{files to appen}

\item{Folder}{Folder where files are, if missing: CachePath will be called}

\item{Name}{Name to save CSV}

\item{Remove}{boolean: if true all files will be deleted after}
}
\description{
This function allows to append a set of csv files into one.
}
\examples{
TODO
}
\seealso{
\code{\link{TAPChunks}}

Other Internal Utilities: \code{\link{CheckChunkInfo}},
  \code{\link{CheckDemogDB}}, \code{\link{CheckMemory}},
  \code{\link{CheckRequirements}},
  \code{\link{CheckSourceYaml}},
  \code{\link{CheckSpaceDisk}}, \code{\link{ConnectToADL}},
  \code{\link{FormatList}}, \code{\link{GetConfiguration}},
  \code{\link{GetMapping}}, \code{\link{HttpCode}},
  \code{\link{LoadDemogDB}}, \code{\link{OptionsMessage}},
  \code{\link{OrderColumn}}, \code{\link{SaveLocal}},
  \code{\link{TestNet}}, \code{\link{TestVPN}},
  \code{\link{TimeStamps}}, \code{\link{Unzip}},
  \code{\link{ValidateChunk}}
}
\author{
Jonathan Tooley Associados Lda
}
