% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Manipulation.R
\name{ApplyLongitudinalSet}
\alias{ApplyLongitudinalSet}
\title{Forming the Longitudinal Data Set}
\usage{
ApplyLongitudinalSet(Data)
}
\arguments{
\item{Data}{TODO}
}
\author{
Jonathan Tooley Associados Lda
}
