% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Manipulation.R
\name{ApplyPenetration}
\alias{ApplyPenetration}
\title{Apply/Calculate penetration in Data Set}
\usage{
ApplyPenetration(Data, Dimension, PenetrationDimension)
}
\arguments{
\item{Data}{Dataset where the penetration will be applied}

\item{Dimension}{Dimensions to be included in the final output}

\item{PenetrationDimension}{Dimension that is the Penetration target}
}
\author{
Jonathan Tooley Associados Lda
}
