% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BlueText.R
\name{BlueText}
\alias{BlueText}
\title{BlueText}
\usage{
BlueText(Text)
}
\arguments{
\item{Text}{String to apply font color}
}
\description{
Add Blue font color
}
\examples{
  cat(BlueText("Hello World!"))
message(BlueText("Status100"))
}
\seealso{
Other Text Formatters: \code{\link{BoldText}},
  \code{\link{GreenText}}, \code{\link{RedText}},
  \code{\link{UnderLineText}}, \code{\link{YellowText}},
  \code{\link{catInfo}}
}
\keyword{internal}
