% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoldText.R
\name{BoldText}
\alias{BoldText}
\title{Printing Bold Text}
\usage{
BoldText(Text)
}
\arguments{
\item{Text}{String to apply bold codes.}
}
\value{
Returns a string
}
\description{
This function bolds the text submitted by wrapping the text
with ANSI escape sequences.
}
\examples{
  cat(TAPChunks:::BoldText("Hello World!"))
message(TAPChunks:::BoldText("Status100"))
}
\seealso{
\code{\link{TAPChunks}}

Other Text Formatters: \code{\link{BlueText}},
  \code{\link{GreenText}}, \code{\link{ItalicText}},
  \code{\link{RedText}}, \code{\link{StrikeText}},
  \code{\link{UnderLineText}}, \code{\link{YellowText}}
}
\author{
JTA - The Data Scientists
}
\concept{Text Formatters}
\keyword{internal}
