% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Configuration.R
\name{catSuccess}
\alias{catSuccess}
\title{catSuccess}
\usage{
catSuccess(Text, Highlight = NULL)
}
\arguments{
\item{Text}{TODO}

\item{Highlight}{TODO}
}
\description{
Print Success message with Highlighed text
}
