% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Configuration.R
\name{catSuperUser}
\alias{catSuperUser}
\title{catSuperUser}
\usage{
catSuperUser(Text, Highlight = NULL)
}
\arguments{
\item{Text}{TODO}

\item{Highlight}{TODO}
}
\description{
Print Super User messages with Highlighed text
}
