% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InternalUtilities.R
\name{CheckChunkInfo}
\alias{CheckChunkInfo}
\title{Check Chunk Files information}
\usage{
CheckChunkInfo(Source = "ALL", From = "", To = "")
}
\arguments{
\item{Source}{TODO}

\item{From}{The timestamp of the first month of data.  If this is not provided this defaults
to 2016M01 but in general the user should select a timestamp.}

\item{To}{The timestamp of the last month of data needed when the user needs to get a range
of timestamps.  If this is left blank then the system returns just the single timestamp provided
in the to parameter.}
}
\description{
TODO
}
\seealso{
\code{\link{TAPChunks}}

Other Internal Utilities: \code{\link{AppendCSV}},
  \code{\link{CheckDemogDB}}, \code{\link{CheckMemory}},
  \code{\link{CheckRequirements}},
  \code{\link{CheckSourceYaml}},
  \code{\link{CheckSpaceDisk}}, \code{\link{ConnectToADL}},
  \code{\link{FormatList}}, \code{\link{GetConfiguration}},
  \code{\link{GetMapping}}, \code{\link{HttpCode}},
  \code{\link{LoadDemogDB}}, \code{\link{OptionsMessage}},
  \code{\link{OrderColumn}}, \code{\link{SaveLocal}},
  \code{\link{TestNet}}, \code{\link{TestVPN}},
  \code{\link{TimeStamps}}, \code{\link{Unzip}},
  \code{\link{ValidateChunk}}
}
\author{
Jonathan Tooley Associados Lda
}
\keyword{internal}
