% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckPackage.R
\name{CheckPackage}
\alias{CheckPackage}
\title{CheckPackage}
\usage{
CheckPackage(lbr)
}
\arguments{
\item{lbr}{Library}
}
\description{
Check if the package exists and install in case doesn't exist.
}
\examples{
CheckPackage(x)
}
\seealso{
\code{\link{TAPChunks}}
}
\author{
JTA - The Data Scientists
}
