% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckSpaceDisk.R
\name{CheckSpaceDisk}
\alias{CheckSpaceDisk}
\title{Check The Space on the User's Disk}
\usage{
CheckSpaceDisk(Request)
}
\arguments{
\item{Request}{chunk files required}
}
\description{
When the TAPChunks system writes a chunk to disk this function is used to see if there is
sufficient space to store the file.
}
\details{
The function makes an operating system call to the WMI (Windows Management Instrumentation)
command line utility to see disk space available.  This is then compared with the expected size of
the chunk.  If there is space available then a success message is printed. If the size of the disk is
close to the requirment a warning message is printed.  When there is no space an error condition is
triggered which will serve a message and halt execution.
}
\examples{
CheckSpaceDisk("SWWorkloadWL2017M12.Rdata")
}
\seealso{
\code{\link{TAPChunks}}

Other Internal Utilities: \code{\link{AddDimension}},
  \code{\link{AddFirmographics}},
  \code{\link{AddHierarchy}}, \code{\link{CheckChunkInfo}},
  \code{\link{CheckDemogDB}}, \code{\link{CheckMemory}},
  \code{\link{CheckPackage}},
  \code{\link{CheckRequirements}},
  \code{\link{CheckSourceYaml}},
  \code{\link{ConnectToADL}}, \code{\link{CreateTable}},
  \code{\link{DetectError}}, \code{\link{FormatList}},
  \code{\link{GetConfiguration}}, \code{\link{GetMapping}},
  \code{\link{GetTimestampByObjInternal}},
  \code{\link{GetTimestampByObj}}, \code{\link{HttpCode}},
  \code{\link{LoadDemogDB}},
  \code{\link{OpenSqlConnection}},
  \code{\link{OptionsMessage}}, \code{\link{OrderColumn}},
  \code{\link{RMarkdown}},
  \code{\link{RecreateReadADLChunkExpression}},
  \code{\link{SaveLocal}}, \code{\link{ShowADLPath}},
  \code{\link{ShowADLPrefix}},
  \code{\link{ShowConfigurationPath}},
  \code{\link{TestNet}}, \code{\link{TestVPN}},
  \code{\link{TimeStamps}}, \code{\link{Unzip}},
  \code{\link{UpdateTAPChunks}},
  \code{\link{ValidateChunk}}, \code{\link{WriteFmt}}
}
\author{
JTA - The Data Scientists
}
\concept{Internal Utilities}
\keyword{internal}
