% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CacheControls.R
\name{CleanMemory}
\alias{CleanMemory}
\title{Clean large tables from the memory}
\usage{
CleanMemory(LimitMB = 10)
}
\arguments{
\item{LimitMB}{Limit value in MB.}
}
\value{
Cleans the memory, no return.
}
\description{
This function removes all the objects that
are using more than the specified number of Mbytes in ram
}
\details{
Per default the is size cut-off is set as 10MB so all
tables above this size will be cleared from memory. The function
will not remove any tables that have had their protect attribut set.
See \code{\link{Protect}} and \code{\link{Unprotect}} for more information
on protecting tables in memory.
}
\examples{
CleanMemory(20)
}
\seealso{
\code{\link{TAPChunks}}

Other Cache Controls: \code{\link{DeleteCache}},
  \code{\link{DeleteOutdatedCache}},
  \code{\link{ListCacheDirectory}}, \code{\link{Protect}},
  \code{\link{RefreshCache}}, \code{\link{ShowCachePath}},
  \code{\link{ShowCacheStatus}},
  \code{\link{ShowConfigurationPath}},
  \code{\link{Unprotect}}
}
\author{
Jonathan Tooley Associados Lda
}
