% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InternalUtilities.R
\name{ConvertTime}
\alias{ConvertTime}
\title{ConvertTime (Internal Function used in ExportToSQL)}
\usage{
ConvertTime(TimeSec)
}
\arguments{
\item{TimeSec}{TODO}
}
\description{
TODO
}
\examples{
TODO
}
\seealso{
\code{\link{TAPChunks}}
}
\author{
Jonathan Tooley Associados Lda
}
