% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CacheControls.R
\name{DeleteOutdatedCache}
\alias{DeleteOutdatedCache}
\title{Delete outdated Rdata files in cache folder}
\usage{
DeleteOutdatedCache()
}
\value{
Cleans old files in cache, no return.
}
\description{
Delete Rdata files with more 30 days in cache folder
}
\examples{
DeleteOutdatedCache()
}
\seealso{
\code{\link{TAPChunks}}

Other Cache Controls: \code{\link{CleanMemory}},
  \code{\link{DeleteCache}},
  \code{\link{ListCacheDirectory}}, \code{\link{Protect}},
  \code{\link{RefreshCache}}, \code{\link{ShowCachePath}},
  \code{\link{ShowCacheStatus}},
  \code{\link{ShowConfigurationPath}},
  \code{\link{Unprotect}}
}
\author{
Jonathan Tooley Associados Lda
}
