% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExportToSQL.R
\name{ExportToSQL}
\alias{ExportToSQL}
\title{Export to SQL}
\usage{
ExportToSQL(Connection, InputData, DbSchema, DbTableName,
  ServerPath = ".", Append = FALSE)
}
\arguments{
\item{Connection}{OpenSqlConnection(Connection, DataBase) TODO}

\item{InputData}{class = Data.frame | Data.table, table in r}

\item{DbSchema}{class = string, name of schema in SQL}

\item{DbTableName}{class = string, name for table in SQl}

\item{ServerPath}{path where the FMT and TSV file}

\item{Append}{default = FALSE, When Append = F it creates a new table. If not, adds new information to an exist table in SQL}
}
\value{
None
}
\description{
This function inserts a R table into SQL
}
\seealso{
Other Reporting and saving tools: \code{\link{AppendCSV}},
  \code{\link{ApplyPenetration}}, \code{\link{PublishADL}},
  \code{\link{SaveADL}}, \code{\link{UseAsSlicer}}
}
\concept{Reporting and saving tools}
