% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Configuration.R
\name{FormatList}
\alias{FormatList}
\title{FormatList}
\usage{
FormatList(ListElement, Title = "Title:", NumberCol = 3)
}
\arguments{
\item{ListElement}{List to be applied}

\item{Title}{List title}

\item{NumberCol}{Number the columns the list will have}
}
\description{
Format the list in columns
}
\seealso{
Other Internal Utilities: \code{\link{AppendCSV}},
  \code{\link{CheckChunkInfo}}, \code{\link{CheckDemogDB}},
  \code{\link{CheckMemory}},
  \code{\link{CheckRequirements}},
  \code{\link{CheckSourceYaml}},
  \code{\link{CheckSpaceDisk}}, \code{\link{ConnectToADL}},
  \code{\link{GetConfiguration}}, \code{\link{GetMapping}},
  \code{\link{HttpCode}}, \code{\link{LoadDemogDB}},
  \code{\link{OptionsMessage}}, \code{\link{OrderColumn}},
  \code{\link{SaveLocal}}, \code{\link{TestNet}},
  \code{\link{TestVPN}}, \code{\link{TimeStamps}},
  \code{\link{Unzip}}, \code{\link{ValidateChunk}}
}
