% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HTTPCode.R
\name{HttpCode}
\alias{HttpCode}
\title{Decode the Http Error codes into messages}
\usage{
HttpCode(ErrorCode = 0)
}
\arguments{
\item{ErrorCode}{The client identifier}
}
\value{
Returns a message as a string
}
\description{
This function decode the Http Error codes into messages
}
\examples{
TAPChunks:::httpCode(200)
}
\seealso{
\code{\link{TAPChunks}}

Other Internal Utilities: \code{\link{AddDimension}},
  \code{\link{AddFirmographics}},
  \code{\link{AddHierarchy}}, \code{\link{CheckChunkInfo}},
  \code{\link{CheckDemogDB}}, \code{\link{CheckMemory}},
  \code{\link{CheckPackage}},
  \code{\link{CheckRequirements}},
  \code{\link{CheckSourceYaml}},
  \code{\link{CheckSpaceDisk}}, \code{\link{ConnectToADL}},
  \code{\link{CreateTable}}, \code{\link{DetectError}},
  \code{\link{FormatList}}, \code{\link{GetConfiguration}},
  \code{\link{GetMapping}},
  \code{\link{GetTimestampByObjInternal}},
  \code{\link{GetTimestampByObj}},
  \code{\link{LoadDemogDB}},
  \code{\link{OpenSqlConnection}},
  \code{\link{OptionsMessage}}, \code{\link{OrderColumn}},
  \code{\link{RMarkdown}},
  \code{\link{RecreateReadADLChunkExpression}},
  \code{\link{SaveLocal}}, \code{\link{ShowADLPath}},
  \code{\link{ShowADLPrefix}},
  \code{\link{ShowConfigurationPath}},
  \code{\link{TestNet}}, \code{\link{TestVPN}},
  \code{\link{TimeStamps}}, \code{\link{Unzip}},
  \code{\link{UpdateTAPChunks}},
  \code{\link{ValidateChunk}}, \code{\link{WriteFmt}}
}
\author{
JTA - The Data Scientists
}
\concept{Internal Utilities}
\keyword{internal}
