% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Manipulation.R
\name{JoinChunks}
\alias{JoinChunks}
\title{Joining two data chunks to form one}
\usage{
JoinChunks(...)
}
\arguments{
\item{i}{A data chunk}

\item{j}{A data chunk}
}
\value{
Returns a data table with the two submitted tables appended together
}
\description{
This function joins two chunks of data together to form one appended table.
}
\details{
This function performs a full join of two data chunks by appending the two chunks
together. The resulting data chunk has all the data from both submitted chunks and
so is a full outer join. If the two submitted chunks have the same column format then
the resulting chunk will also retain this format. If either of the submitted chunks
has a column not present in the other then this column will be retained in the output
but.  In order to do this the missing column is created in the input chunk that doesn't
have the column.  This created chunk is filled with the value NA.\cr\cr
The data is not reweighted after join.
}
\seealso{
\code{\link{TAPChunks}}

Other Chunk Manipulators: \code{\link{AddFirmographics}},
  \code{\link{AddOSHierarchy}},
  \code{\link{AggregateChunk}},
  \code{\link{CalculateColumn}}, \code{\link{FilterChunk}},
  \code{\link{RMarkdown}}, \code{\link{ShapeChunk}},
  \code{\link{UseAsSlicer}}
}
\author{
Jonathan Tooley Associados Lda
}
