% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ListADLDirectory.R
\name{ListADLDirectory}
\alias{ListADLDirectory}
\title{List files in the Data Lake}
\usage{
ListADLDirectory(Directory = "", Info = F)
}
\arguments{
\item{Directory}{Specific directory in an Azure Data Lake}

\item{Info}{if true return the modification time}
}
\value{
Returns a data table with a list of available filenames or the message off-line
if the user does not have net connectivity.
}
\description{
A function to return a directory of files in the data lake for the specificied source.
}
\details{
Telemetry data is stored in an Azure Data Lake and is organized by data source.
This function allows the user to retrieve a directory listing of all available files
relating to a particular data source.
}
\examples{
ListADLDirectory("Email")
}
\seealso{
\code{\link{TAPChunks}}

Other Data Lake Tools: \code{\link{ReadABSFile}},
  \code{\link{ReadADLChunk}}, \code{\link{ReadADLFile}},
  \code{\link{UploadLocallyToADL}}
}
\author{
JTA - The Data Scientists
}
\concept{Data Lake Tools}
