% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ListCacheDirectory.R
\name{ListCacheDirectory}
\alias{ListCacheDirectory}
\title{List files in the local cache}
\usage{
ListCacheDirectory(Source = "", ShowAttributes = FALSE)
}
\arguments{
\item{Source}{If the Source parameter is not provided then the interactive mode
requests a numeric selection (n). A user may also supply a string
with one of the following (Source) values:\cr
\tabular{rlll}{
\tab  n \strong{Source}       \tab \strong{Effect} \tab \strong{Data Documentation:}\cr
\tab  1 \emph{Email}          \tab For the Spiceworks Email data \tab \code{\link{TestEmailChunk}}  \cr
\tab  2 \emph{Client}         \tab For the Spiceworks PC Client data \tab \code{\link{TestClientChunk}} \cr
\tab  3 \emph{Networkhv}      \tab For the Spiceworks Network data by hypervisor \tab \code{\link{TestNetworkhvChunk}} \cr
\tab  4 \emph{Networkos}      \tab For the Spiceworks Network data by operating system \tab \code{\link{TestNetworkosChunk}} \cr
\tab  5 \emph{ServerRoles}    \tab For the Spiceworks Server Roles data \tab \code{\link{TestServerRolesChunk}} \cr
\tab  6 \emph{WorkloadDB}     \tab For the Spiceworks Workload data for database products \tab \code{\link{TestWorkloadDBChunk}} \cr
\tab  7 \emph{WorkloadEco}    \tab This chunk lists the virt flag and data type along with Hypervisor and OS for the intersection of Network and VID \tab \code{\link{TestWorkloadEcoChunk}} \cr
\tab  8 \emph{WorkloadHW}     \tab This chunk lists the metadata that we received relating to server hardware. \tab \code{\link{TestWorkloadHWChunk}} \cr
\tab  9 \emph{WorkloadN}      \tab This chunk lists the virt flag and data type along with Hypervisor and OS for the Network style data \tab \code{\link{TestWorkloadNChunk}} \cr
\tab 10 \emph{WorkloadVID}    \tab This chunk lists the virt flag and data type along with Hypervisor and OS for the VID style data \tab \code{\link{TestWorkloadVIDChunk}} \cr
\tab 11 \emph{WorkloadWL}     \tab For the Spiceworks workload data for non-database products \tab \code{\link{TestWorkloadWLChunk}} \cr
\tab 12 \emph{CCMCost}        \tab For the Spiceworks CCM Cost data\cr
\tab 13 \emph{CCMUsage}       \tab For the Spiceworks CCM Usage data\cr
\tab 14 \emph{OnlineServices} \tab For the Spiceworks Online Services Data\cr
\tab 15 \emph{VID}            \tab For the Spiceworks VID Data\cr
\tab 16 \emph{ServerAge}      \tab For the Spiceworks Server Age Data\cr
\tab 17 \emph{Cosmos}         \tab For the Cosmos Data\cr
}}

\item{ShowAttributes}{When TRUE a detailed listing of the files attributes is shown.}
}
\value{
Returns a data table with a columns called 'File Name' holding strings with
the name of each file found.  If the user selected ShowAttributes = TRUE then the data
table will have additional columns for 'File Size' and 'File Created'
}
\description{
Show a list of files that we have in the local cache
}
\details{
This function will establish if the user has a local cache directory installed
in their machine.  If it doesn't exist it will be automatically created.
The function then returns a data table that has a list of chunks that are stored
in the cache.  \cr\cr
The user may select an optional filter using the source parameter
that will filter the results to show only chunks that relate to the source.\cr\cr
If the user selects showattributes = TRUE then the returned data table has
a series of properties listed that will help the user understand the formation of the
chunk. \cr
}
\examples{
ListCacheDirectory()
ListCacheDirectory(ShowAttributes = TRUE)
ListCacheDirectory("Client", ShowAttributes = TRUE)
}
\seealso{
\code{\link{TAPChunks}}

Other Cache Controls: \code{\link{DeleteCache}},
  \code{\link{DeleteOutdatedCache}},
  \code{\link{RefreshCache}}, \code{\link{ShowCachePath}},
  \code{\link{ShowCacheStatus}}, \code{\link{Unprotect}}
}
\author{
JTA - The Data Scientists
}
\concept{Cache Controls}
