% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpenSqlConnection.R
\name{OpenSqlConnection}
\alias{OpenSqlConnection}
\title{Open SQL Connection}
\usage{
OpenSqlConnection(Server, Database, Uid = NULL, Pwd = NULL)
}
\arguments{
\item{Server}{name of server SQL}

\item{Database}{name of database SQL}

\item{Uid}{user identification of server SQL}

\item{Pwd}{password of server SQL}
}
\value{
Connection with SQL
}
\description{
This function can open connection with SQL without create a odbc connection
}
\seealso{
Other Internal Utilities: \code{\link{AddDimension}},
  \code{\link{AddFirmographics}},
  \code{\link{AddHierarchy}}, \code{\link{CheckChunkInfo}},
  \code{\link{CheckDemogDB}}, \code{\link{CheckMemory}},
  \code{\link{CheckPackage}},
  \code{\link{CheckRequirements}},
  \code{\link{CheckSourceYaml}},
  \code{\link{CheckSpaceDisk}}, \code{\link{ConnectToADL}},
  \code{\link{CreateTable}}, \code{\link{DetectError}},
  \code{\link{FormatList}}, \code{\link{GetConfiguration}},
  \code{\link{GetMapping}},
  \code{\link{GetTimestampByObjInternal}},
  \code{\link{GetTimestampByObj}}, \code{\link{HttpCode}},
  \code{\link{LoadDemogDB}}, \code{\link{OptionsMessage}},
  \code{\link{OrderColumn}}, \code{\link{RMarkdown}},
  \code{\link{RecreateReadADLChunkExpression}},
  \code{\link{SaveLocal}}, \code{\link{ShowADLPath}},
  \code{\link{ShowADLPrefix}},
  \code{\link{ShowConfigurationPath}},
  \code{\link{TestNet}}, \code{\link{TestVPN}},
  \code{\link{TimeStamps}}, \code{\link{Unzip}},
  \code{\link{UpdateTAPChunks}},
  \code{\link{ValidateChunk}}, \code{\link{WriteFmt}}
}
\author{
JTA - The Data Scientists
}
\concept{Internal Utilities}
\keyword{internal}
