% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Configuration.R
\name{OptionsMessage}
\alias{OptionsMessage}
\title{OptionsMessage}
\usage{
OptionsMessage(Error = T, Warning = T, Info = T, Success = T)
}
\arguments{
\item{Error}{Want to get Error messages?}

\item{Warning}{Want to get warning messages?}

\item{Info}{Want to get info messages?}

\item{Success}{Want to get success messages?}
}
\description{
To define which message type the users want to see.
}
\seealso{
Other Internal Utilities: \code{\link{AppendCSV}},
  \code{\link{CheckChunkInfo}}, \code{\link{CheckDemogDB}},
  \code{\link{CheckMemory}},
  \code{\link{CheckRequirements}},
  \code{\link{CheckSourceYaml}},
  \code{\link{CheckSpaceDisk}}, \code{\link{ConnectToADL}},
  \code{\link{FormatList}}, \code{\link{GetConfiguration}},
  \code{\link{GetMapping}}, \code{\link{HttpCode}},
  \code{\link{LoadDemogDB}}, \code{\link{OrderColumn}},
  \code{\link{SaveLocal}}, \code{\link{TestNet}},
  \code{\link{TestVPN}}, \code{\link{TimeStamps}},
  \code{\link{Unzip}}, \code{\link{ValidateChunk}}
}
