% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Protect.R
\name{Protect}
\alias{Protect}
\title{Protect a table in memory from deletion}
\usage{
Protect(...)
}
\arguments{
\item{...}{A list of objects that the user wishes to protect or to have created as empty protected chunks.}
}
\value{
A object with attribute set as protect.
}
\description{
This function has two main uses:
\enumerate{
\item Protect an existing chunk from being cleaned from memory
\item Create an empty protected chunk}
adds an attribute to a TAP Chunk that protects the object from being deleted
when the user uses the CleanMemory function.
}
\details{
The function receives a list of objects.  If the object already exists
then it will be protected.  If the object is not defined then it will be
created and protected.
}
\seealso{
\code{\link{TAPChunks}}

Other Compute Tools: \code{\link{BeginIterate}},
  \code{\link{CleanMemory}}, \code{\link{EndIterate}},
  \code{\link{IterateScript}}
}
\author{
JTA - The Data Scientists
}
\concept{Compute Tools}
