% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PublishADL.R
\name{PublishADL}
\alias{PublishADL}
\title{Publish File into the Data Lake}
\usage{
PublishADL(Data, Project, ChunkName, FileExtension = "csv",
  Replace = F)
}
\arguments{
\item{Data}{A TAP Chunk to be saved to the data lake}

\item{Project}{Name of the project. If the project doesn't exist a new folder will be created in the ADL.}

\item{ChunkName}{The name that is given to the data on the lake.}

\item{FileExtension}{A string containing either 'csv' or 'rdata' used to select the format in which the file is saved.}

\item{Replace}{Replace in case exists a file with same name. By default is FALSE.}
}
\description{
A function to save files in the Publication folder in the data lake. This function is similar to SaveADL() function, the difference is the chunk is saved in the Publish area.
By default saves a CSV file but also supports Rdata files.
}
\examples{
PublishADL(TestEmailChunk, Project="Emailtracker", ChunkName = "FactFile", FileExtension="csv", Replace = T)
}
\seealso{
\code{\link{TAPChunks}}

Other Reporting and saving tools: \code{\link{AppendCSV}},
  \code{\link{ApplyPenetration}},
  \code{\link{ExportToSQL}}, \code{\link{SaveADL}},
  \code{\link{UseAsSlicer}}
}
\author{
JTA - The Data Scientists
}
\concept{Reporting and saving tools}
