% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadADLFile.R
\name{ReadADLFile}
\alias{ReadADLFile}
\title{Reading user data from the Data Lake}
\usage{
ReadADLFile(SpecificPath, FileName, SaveChunk = T, ShowInfo = T)
}
\arguments{
\item{SpecificPath}{Specific folder path where the file is located}

\item{FileName}{Name of the file}

\item{SaveChunk}{Save chunk in your local machine}

\item{ShowInfo}{Show extra information}
}
\value{
Prints a status message.
}
\description{
This function allows users to read in a file that has been previously
stored on the Azure Data Lake. Chunks of telemetry data may be more easily read by
using \code{\link{ReadADLChunk}} which simplifies the process for pre-prepared chunks.
This function allows users more flexibility to select the file they wish and they may
also get added file metadata too.
}
\examples{
TODO
}
\seealso{
\code{\link{TAPChunks}}

Other Data Lake Tools: \code{\link{ListADLDirectory}},
  \code{\link{ReadABSFile}}, \code{\link{ReadADLChunk}},
  \code{\link{UploadLocallyToADL}}
}
\author{
JTA - The Data Scientists
}
\concept{Data Lake Tools}
