% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RedText.R
\name{RedText}
\alias{RedText}
\title{RedText}
\usage{
RedText(Text)
}
\arguments{
\item{Text}{String to add font color}
}
\description{
Apply Red font color
}
\examples{
  cat(RedText("Hello World!"))
message(RedText("Status100"))
}
\seealso{
Other Text Formatters: \code{\link{BlueText}},
  \code{\link{BoldText}}, \code{\link{GreenText}},
  \code{\link{UnderLineText}}, \code{\link{YellowText}},
  \code{\link{catInfo}}
}
\keyword{internal}
