% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaveLocal.R
\name{SaveLocal}
\alias{SaveLocal}
\title{Save Objects}
\usage{
SaveLocal(..., FileExtension = "csv", Path = NULL)
}
\arguments{
\item{...}{Objects to be saved}

\item{FileExtension}{A string with the required file extension. This can be one of two types, csv or rdata.}

\item{Path}{Destination of the files with information of the R session objects}
}
\description{
This function allows to save R environment objects into local machine. Formats available csv and rdata
}
\examples{
TAPChunks:::SaveLocal(TestClientChunk, TestEmailChunk, path = getwd());
TAPChunks:::SaveLocal("TestClientChunk", "TestEmailChunk", FileExtension = "rdata", path = getwd())
}
\seealso{
\code{\link{TAPChunks}}

Other Internal Utilities: \code{\link{AddDimension}},
  \code{\link{AddFirmographics}},
  \code{\link{AddHierarchy}}, \code{\link{CheckChunkInfo}},
  \code{\link{CheckDemogDB}}, \code{\link{CheckMemory}},
  \code{\link{CheckPackage}},
  \code{\link{CheckRequirements}},
  \code{\link{CheckSourceYaml}},
  \code{\link{CheckSpaceDisk}}, \code{\link{ConnectToADL}},
  \code{\link{CreateTable}}, \code{\link{DetectError}},
  \code{\link{FormatList}}, \code{\link{GetConfiguration}},
  \code{\link{GetMapping}},
  \code{\link{GetTimestampByObjInternal}},
  \code{\link{GetTimestampByObj}}, \code{\link{HttpCode}},
  \code{\link{LoadDemogDB}},
  \code{\link{OpenSqlConnection}},
  \code{\link{OptionsMessage}}, \code{\link{OrderColumn}},
  \code{\link{RMarkdown}},
  \code{\link{RecreateReadADLChunkExpression}},
  \code{\link{ShowADLPath}}, \code{\link{ShowADLPrefix}},
  \code{\link{ShowConfigurationPath}},
  \code{\link{TestNet}}, \code{\link{TestVPN}},
  \code{\link{TimeStamps}}, \code{\link{Unzip}},
  \code{\link{UpdateTAPChunks}},
  \code{\link{ValidateChunk}}, \code{\link{WriteFmt}}
}
\author{
JTA - The Data Scientists
}
\concept{Internal Utilities}
\keyword{internal}
