% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Manipulation.R
\name{SelectColumns}
\alias{SelectColumns}
\title{Select Columns in Chunk}
\usage{
SelectColumns(Source = NULL, Columns = NULL)
}
\arguments{
\item{Source}{chunk}

\item{Columns}{List of name columns to extract}
}
\description{
This function allows to select which columns from the chunk to extract
}
\examples{
SelectColumns(Source = Chunk, Columns = c("uuid", "variable", "value"))
}
\seealso{
\code{\link{TAPChunks}}
}
\author{
Jonathan Tooley Associados Lda
}
