% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Manipulation.R
\name{ShapeChunk}
\alias{ShapeChunk}
\title{Shape chunk - Pivot and Unpivot}
\usage{
ShapeChunk(Data, Variable = NULL, Value = NULL, Horizontal = T,
  Operation = "default")
}
\arguments{
\item{Data}{A data table}

\item{Variable}{The column name to reshape}

\item{Value}{The column name to fill the columns}

\item{Horizontal}{Boolean: TRUE as default}
}
\value{
Returns a data table with shaped as user requested
}
\description{
This function will reshape a data table, it will pivot a variable to several columns.
}
\details{
This function takes a data table, a variable and value to reshape. The unique values
of the variable will become new columns and the value will fill those columns. NAs will
be replaced by 0.
}
\seealso{
\code{\link{TAPChunks}}
Issue #162

Other Chunk Manipulators: \code{\link{AddFirmographics}},
  \code{\link{AddOSHierarchy}},
  \code{\link{AggregateChunk}},
  \code{\link{CalculateColumn}}, \code{\link{FilterChunk}},
  \code{\link{JoinChunks}}, \code{\link{RMarkdown}},
  \code{\link{UseAsSlicer}}
}
\author{
Jonathan Tooley Associados Lda
}
