% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShowADLPrefix.R
\name{ShowADLPrefix}
\alias{ShowADLPrefix}
\title{Find the filename prefix associated with a source}
\usage{
ShowADLPrefix(Source = "")
}
\arguments{
\item{Source}{Name of the source, if the Source parameter is not provided then a list of sources will be shown:\cr
\tabular{lll}{
\tab \strong{Source}       \tab \strong{Effect} \tab \strong{Data Documentation:}\cr
\tab \emph{Email}          \tab For the Spiceworks Email data \tab \code{\link{TestEmailChunk}}  \cr
\tab \emph{Client}         \tab For the Spiceworks PC Client data \tab \code{\link{TestClientChunk}} \cr
\tab \emph{Networkhv}      \tab For the Spiceworks Network data by hypervisor \tab \code{\link{TestNetworkhvChunk}} \cr
\tab \emph{Networkos}      \tab For the Spiceworks Network data by operating system \tab \code{\link{TestNetworkosChunk}} \cr
\tab \emph{ServerRoles}    \tab For the Spiceworks Server Roles data \tab \code{\link{TestServerRolesChunk}} \cr
\tab \emph{WorkloadDB}     \tab For the Spiceworks Workload data for database products \tab \code{\link{TestWorkloadDBChunk}} \cr
\tab \emph{WorkloadEco}    \tab This chunk lists the virt flag and data type along with Hypervisor and OS for the intersection of Network and VID \tab \code{\link{TestWorkloadEcoChunk}} \cr
\tab \emph{WorkloadHW}     \tab This chunk lists the metadata that we received relating to server hardware. \tab \code{\link{TestWorkloadHWChunk}} \cr
\tab \emph{WorkloadN}      \tab This chunk lists the virt flag and data type along with Hypervisor and OS for the Network style data \tab \code{\link{TestWorkloadNChunk}} \cr
\tab \emph{WorkloadVID}    \tab This chunk lists the virt flag and data type along with Hypervisor and OS for the VID style data \tab \code{\link{TestWorkloadVIDChunk}} \cr
\tab \emph{WorkloadWL}     \tab For the Spiceworks workload data for non-database products \tab \code{\link{TestWorkloadWLChunk}} \cr
\tab \emph{CCMCost}        \tab For the Spiceworks CCM Cost data\cr
\tab \emph{CCMUsage}       \tab For the Spiceworks CCM Usage data\cr
\tab \emph{OnlineServices} \tab For the Spiceworks Online Services Data\cr
\tab \emph{VID}            \tab For the Spiceworks VID Data\cr
\tab \emph{ServerAge}      \tab For the Spiceworks Server Age Data\cr
\tab \emph{Cosmos}         \tab For the Cosmos Data\cr
}}
}
\value{
Returns a string with the prefix for the given source
}
\description{
The TAP Data Lake stores files with a unique filename.  The
name consists of a prefix (that depends on the source) and the timestamp.
This function tells you the prefix text that the system uses for the selected source.
}
\examples{
ShowADLPrefix("Email")
}
\seealso{
\code{\link{TAPChunks}}

Other Data Lake Controls: \code{\link{ReadADLChunk}},
  \code{\link{ShowADLPath}}
}
\author{
JTA - The Data Scientists
}
\keyword{internal}
