% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CacheControls.R
\name{ShowCachePath}
\alias{ShowCachePath}
\title{Establishing, pointing and set to the cache folder}
\usage{
ShowCachePath(Directory)
}
\arguments{
\item{Directory}{Path to set the Cache location}
}
\value{
A string with the path to the cache folder.
}
\description{
This function allocates a path where the cache will be stored, if Directory parameter is NULL then it will assume a location in user home directory. 
It then checks that the cache exists, creating it if not.
}
\details{
If the cache is placed in the users home directory in a folder called
TAPCache.
}
\examples{
ShowCachePath()
}
\seealso{
\code{\link{TAPChunks}}

Other Cache Controls: \code{\link{CleanMemory}},
  \code{\link{DeleteCache}},
  \code{\link{DeleteOutdatedCache}},
  \code{\link{ListCacheDirectory}}, \code{\link{Protect}},
  \code{\link{RefreshCache}},
  \code{\link{ShowCacheStatus}},
  \code{\link{ShowConfigurationPath}},
  \code{\link{Unprotect}}
}
\author{
Jonathan Tooley Associados Lda
}
