% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShowCacheStatus.R
\name{ShowCacheStatus}
\alias{ShowCacheStatus}
\title{Detecting when data on the Data Lake has been updated}
\usage{
ShowCacheStatus(Source = "")
}
\arguments{
\item{Source}{Name of the source, if the Source parameter is not provided then a list of sources will be shown:\cr
\tabular{lll}{
\tab \strong{Source}       \tab \strong{Effect} \tab \strong{Data Documentation:}\cr
\tab \emph{Email}          \tab For the Spiceworks Email data \tab \code{\link{TestEmailChunk}}  \cr
\tab \emph{Client}         \tab For the Spiceworks PC Client data \tab \code{\link{TestClientChunk}} \cr
\tab \emph{Networkhv}      \tab For the Spiceworks Network data by hypervisor \tab \code{\link{TestNetworkhvChunk}} \cr
\tab \emph{Networkos}      \tab For the Spiceworks Network data by operating system \tab \code{\link{TestNetworkosChunk}} \cr
\tab \emph{ServerRoles}    \tab For the Spiceworks Server Roles data \tab \code{\link{TestServerRolesChunk}} \cr
\tab \emph{WorkloadDB}     \tab For the Spiceworks Workload data for database products \tab \code{\link{TestWorkloadDBChunk}} \cr
\tab \emph{WorkloadEco}    \tab This chunk lists the virt flag and data type along with Hypervisor and OS for the intersection of Network and VID \tab \code{\link{TestWorkloadEcoChunk}} \cr
\tab \emph{WorkloadHW}     \tab This chunk lists the metadata that we received relating to server hardware. \tab \code{\link{TestWorkloadHWChunk}} \cr
\tab \emph{WorkloadN}      \tab This chunk lists the virt flag and data type along with Hypervisor and OS for the Network style data \tab \code{\link{TestWorkloadNChunk}} \cr
\tab \emph{WorkloadVID}    \tab This chunk lists the virt flag and data type along with Hypervisor and OS for the VID style data \tab \code{\link{TestWorkloadVIDChunk}} \cr
\tab \emph{WorkloadWL}     \tab For the Spiceworks workload data for non-database products \tab \code{\link{TestWorkloadWLChunk}} \cr
\tab \emph{CCMCost}        \tab For the Spiceworks CCM Cost data\cr
\tab \emph{CCMUsage}       \tab For the Spiceworks CCM Usage data\cr
\tab \emph{OnlineServices} \tab For the Spiceworks Online Services Data\cr
\tab \emph{VID}            \tab For the Spiceworks VID Data\cr
\tab \emph{ServerAge}      \tab For the Spiceworks Server Age Data\cr
\tab \emph{Cosmos}         \tab For the Cosmos Data\cr
}}
}
\value{
Returns an error status to show if the command completed succesfully or not
}
\description{
Compare the local files in the cache with the files on the
Data Lake and warn the user if any have been superseded.
}
\details{
There are often times when our data provider reissues files after discovering
errors in the feed.  This then requires us to re process the loading of that data.
The issue is that, for the convenience of our users, we allow the local system to
keep a cached copy of the original chunk.  This command allows the user to detect
outdated files. \cr \cr
The data lake is separated into distinct areas for each data source.  Because of this
the user must specify the source for which they wish to get a status update.
}
\examples{
ShowCacheStatus("Email")
}
\seealso{
\code{\link{TAPChunks}}

Other Cache Controls: \code{\link{CleanMemory}},
  \code{\link{DeleteCache}},
  \code{\link{DeleteOutdatedCache}},
  \code{\link{ListCacheDirectory}}, \code{\link{Protect}},
  \code{\link{RefreshCache}}, \code{\link{ShowCachePath}},
  \code{\link{ShowConfigurationPath}},
  \code{\link{Unprotect}}
}
\author{
JTA - The Data Scientists
}
