% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Configuration.R
\name{ShowConfigurationPath}
\alias{ShowConfigurationPath}
\title{Establishing and pointing to the configuration settings}
\usage{
ShowConfigurationPath(Directory)
}
\arguments{
\item{Directory}{TODO}
}
\description{
TODO
}
\examples{
ShowConfigurationPath()
}
\seealso{
\code{\link{TAPChunks}}

Other Cache Controls: \code{\link{CleanMemory}},
  \code{\link{DeleteCache}},
  \code{\link{DeleteOutdatedCache}},
  \code{\link{ListCacheDirectory}}, \code{\link{Protect}},
  \code{\link{RefreshCache}}, \code{\link{ShowCachePath}},
  \code{\link{ShowCacheStatus}}, \code{\link{Unprotect}}
}
\author{
Jonathan Tooley Associados Lda
}
