% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShowLongitudinalSize.R
\name{ShowLongitudinalSize}
\alias{ShowLongitudinalSize}
\title{Show the Size of a Longitudinal Set}
\usage{
ShowLongitudinalSize(Data)
}
\arguments{
\item{Data}{A data chunk that must have an org identifier and a timestamp column.}
}
\description{
This function considers each organisation identifier that exists in
the submitted chunk and looks to see the number of distinct time periods in which
each organization appears. The function then reports the size of the logitudinal
cohort for both Clean and not Clean orgs.
}
\examples{
ShowLongitudinalSize(TestEmailChunk)
}
\seealso{
\code{\link{TAPChunks}}

Other Weighting and cleaning functions: \code{\link{ApplyLongitudinalSet}},
  \code{\link{ApplyWeightCalc}}, \code{\link{CleanChunk}},
  \code{\link{ShowWeightCalc}}
}
\author{
JTA - The Data Scientists
}
\concept{Weighting and cleaning functions}
