% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShowSampleSize.R
\name{ShowSampleSize}
\alias{ShowSampleSize}
\title{Summarizing a file to see the sample size}
\usage{
ShowSampleSize(Data, By = NULL)
}
\arguments{
\item{Data}{A TAP Chunk to be summarised by the function.}

\item{By}{The name of a column in the chunk by which to segment the sample size.}
}
\description{
This function will tell the user the sample size in a file.
The sample size is defined as the number of distinct orgs in the data and is
split between the number of unclean orgs, unweighted clean orgs and weighted clean
orgs.
}
\examples{
ShowSampleSize(TestEmailChunk)
ShowSampleSize(ApplyWeightCalc(TestEmailChunk))
ShowSampleSize(ApplyWeightCalc(TestEmailChunk), By = "email_delivery")
}
\seealso{
Other Chunk Manipulators: \code{\link{AggregateChunk}},
  \code{\link{CalculateColumn}}, \code{\link{FilterChunk}},
  \code{\link{JoinChunks}}, \code{\link{MergeChunks}},
  \code{\link{SelectColumns}}, \code{\link{ShapeChunk}}
}
\author{
JTA - The Data Scientists
}
\concept{Chunk Manipulators}
