% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnderLineText.R
\name{UnderLineText}
\alias{UnderLineText}
\title{UnderLineText}
\usage{
UnderLineText(Text = "")
}
\arguments{
\item{Text}{String to apply underline codes}
}
\value{
Returns a string preceded by Esc[4m and closed with Esc[24m
}
\description{
Apply underline ANSI codes arount the submitted string
}
\examples{
  cat(UnderLineText("Hello World!"))
message(UnderLineText("Status 100"))
}
\seealso{
Other Text Formatters: \code{\link{BlueText}},
  \code{\link{BoldText}}, \code{\link{GreenText}},
  \code{\link{RedText}}, \code{\link{YellowText}},
  \code{\link{catInfo}}
}
\author{
JTA - The Data Scientists
}
\keyword{internal}
