% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Unprotect.R
\name{Unprotect}
\alias{Unprotect}
\title{Unprotect the object that has been assigned as Protected}
\usage{
Unprotect(...)
}
\arguments{
\item{...}{A series of object names to have the protection removed.}
}
\value{
A object without attribute set as protect.
}
\description{
This function removes the Protect attribute from the list of objects passed in the parameter.
When the user uses the CleanMemory function the object will be deleted.
}
\seealso{
\code{\link{TAPChunks}}

Other Cache Controls: \code{\link{DeleteCache}},
  \code{\link{DeleteOutdatedCache}},
  \code{\link{ListCacheDirectory}},
  \code{\link{RefreshCache}}, \code{\link{ShowCachePath}},
  \code{\link{ShowCacheStatus}}
}
\author{
JTA - The Data Scientists
}
\concept{Cache Controls}
