% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UseAsSlicer.R
\name{UseAsSlicer}
\alias{UseAsSlicer}
\title{UseAsSlicer}
\usage{
UseAsSlicer(Data, ReportName = "test")
}
\arguments{
\item{Data}{A data table}

\item{ReportName}{TODO}
}
\value{
TODO
}
\description{
TODO
}
\details{
TODO
}
\seealso{
\code{\link{TAPChunks}}

Other Reporting and saving tools: \code{\link{AppendCSV}},
  \code{\link{ApplyPenetration}},
  \code{\link{ExportToSQL}}, \code{\link{PublishADL}},
  \code{\link{SaveADL}}
}
\author{
JTA - The Data Scientists
}
\concept{Reporting and saving tools}
